### 2025-09-10 #8968 @editorialbot
[PRE REVIEW]: 'torch_bayesian: A Flexible Python Package for Bayesian Neural Networks in PyTorch'
**Submitting author:** @ArvidWeyrauch (<a href="http://orcid.org/0000-0002-2684-0927">Arvid Weyrauch</a>)
**Repository:** https://github.com/RAI-SCC/torch_bayesian
**Branch with paper.md** (empty if default branch): dev
**Version:** v1.0.0
