### 2025-09-21 2fcac75
Add new distribution base and merged Normal
