+++ torch_bayesian/vi/distributions/base.py
@@ -1,0 +1,297 @@
+import math
+from inspect import signature
+from typing import TYPE_CHECKING, Callable, Dict, Tuple, Union
+from warnings import warn
+
+from torch import Tensor
+from torch.nn import init
+
+from ..utils import PostInitCallMeta
+
+if TYPE_CHECKING:
+    from ..base import VIModule  # pragma: no cover
+
+
+class Distribution(metaclass=PostInitCallMeta):
+    """Base class for distributions."""
+
+    distribution_parameters: Tuple[str, ...]
+    _default_variational_parameters: Tuple[float, ...]
+    _required_parameters: Tuple[str, ...] = ()
+    _scaling_parameters: Tuple[str, ...]
+    _rescaled: bool = False
+    sample: Callable[..., Tensor]
+    prior_log_prob: Callable[..., Tensor]
+    variational_log_prob: Callable[..., Tensor]
+    predictive_parameters_from_samples: Callable[
+        [Tensor], Union[Tensor, Tuple[Tensor, ...]]
+    ]
+    log_prob_from_parameters: Callable[
+        [Tensor, Union[Tensor, Tuple[Tensor, ...]]], Tensor
+    ]
+    is_prior: bool = False
+    is_variational_distribution: bool = False
+    is_predictive_distribution: bool = False
+
+    def reset_variational_parameters(
+        self,
+        module: "VIModule",
+        variable: str,
+        fan_in: int,
+        kaiming_scaling: bool = True,
+    ) -> None:
+        """
+        Reset the variational parameters of module.
+
+        Parameters equivalent to non-Bayesian weights (currently "mean", "mode", or
+        "loc") are reset accordingly using Kaiming uniform initialization based on
+        `fan_in` (cf. :meth:`torch.init._calculate_fan_in_and_fan_out`). Other
+        parameters are initialized to the fixed values specified by class defaults,
+        i.e., :attr:`_default_variational_parameters`.
+        If `kaiming_scaling` is ``True`` , the defaults are scaled with
+        `scale` * `default`. Any parameter beginning with "log" is assumed to be in log
+        space and scaled with `default` + log(`scale`). The scale is 1 / sqrt(`fan_in`)
+        for vectors and 1 / sqrt(3 * `fan_in`) for matrices.
+
+        Parameters
+        ----------
+        module: VIModule
+            Module to reset parameters.
+        variable: str
+            Name of the variable to reset.
+        fan_in: int
+            Size of the input parameter map.
+        kaiming_scaling: bool, default: True
+            Whether th scale all parameters according to input map size.
+        """
+        for parameter, default in zip(
+            self.distribution_parameters, self._default_variational_parameters
+        ):
+            parameter_name = module.variational_parameter_name(variable, parameter)
+            param = getattr(module, parameter_name)
+
+            if parameter in ["mean", "mode", "loc"]:
+                self._init_uniform(param, fan_in)
+            elif not kaiming_scaling:
+                init.constant_(param, default)
+            else:
+                is_log = parameter.startswith("log")
+                self._init_constant(param, default, fan_in, is_log)
+
+    def reset_parameters_to_prior(self, module: "VIModule", variable: str) -> None:
+        """
+        Initialize the parameters of a VIModule according to the prior distribution.
+
+        To enable the ``prior_initialization`` functionality, the class must implement
+        this method. It initializes the parameters for one random variable of a
+        :class:`~torch_bayesian.vi.VIModule`, whose variational parameters are
+        supported by the prior, to the prior values. To that end the name of the random
+        variable to initialize is passed to the method.
+
+        This method is called separately for each submodule and therefore does not have
+        to consider any further submodules. It is also called separately for each
+        random variable but should manage all variational parameters the prior can
+        provide. By convention, a prior whose distribution parameters are a true subset
[…]
