+++ torch_bayesian/vi/distributions/__init__.py
@@ -1,0 +1,12 @@
+"""Provides a collection of distributions."""
+
+from .base import Distribution
+from .normal import MeanFieldNormal
+
+Normal = MeanFieldNormal
+
+__all__ = [
+    "Distribution",
+    "MeanFieldNormal",
+    "Normal",
+]
