### 2025-09-21 84e7bd4
Add non-Bayesian pseudo-distribution
