+++ torch_bayesian/vi/distributions/__init__.py
@@ -1,6 +1,7 @@
 """Provides a collection of distributions."""
 
 from .base import Distribution
+from .non_bayesian import NonBayesian, UniformPrior
 from .normal import MeanFieldNormal
 
 Normal = MeanFieldNormal
@@ -8,5 +9,7 @@ Normal = MeanFieldNormal
 __all__ = [
     "Distribution",
     "MeanFieldNormal",
+    "NonBayesian",
     "Normal",
+    "UniformPrior",
 ]
