+++ torch_bayesian/vi/analytical_kl_loss.py
@@ -9,13 +9,8 @@ from torch.nn import Module
 
 from . import _globals
 from .base import VIModule
-from .predictive_distributions import PredictiveDistribution
-from .priors import MeanFieldNormalPrior, Prior, UniformPrior
-from .variational_distributions import (
-    MeanFieldNormalVarDist,
-    NonBayesian,
-    VariationalDistribution,
-)
+from .distributions import Distribution, MeanFieldNormal, NonBayesian
+from .utils import UnsupportedDistributionError
 
 
 def _forward_unimplemented(
@@ -151,7 +146,9 @@ class UniformNormalDivergence(KullbackLeiblerModule):
     """
 
     @staticmethod
-    def forward(variational_mean: Tensor, variational_log_std: Tensor) -> Tensor:
+    def forward(
+        prior_mean: None, variational_mean: Tensor, variational_log_std: Tensor
+    ) -> Tensor:
         """
         Calculate the Kullback-Leibler divergence.
 
@@ -162,6 +159,8 @@ class UniformNormalDivergence(KullbackLeiblerModule):
 
         Parameters
         ----------
+        prior_mean: None
+            Always None, exists for technical reasons.
         variational_mean: Tensor
             Means of the variational distribution.
         variational_log_std: Tensor
@@ -216,12 +215,10 @@ class AnalyticalKullbackLeiblerLoss(Module):
     ----------
     model: :class:`~.VIModule`
         The model to be trained.
-    predictive_distribution: :class:`~.predictive_distributions.PredictiveDistribution`
+    predictive_distribution: :class:`~.distributions.Distribution`
         The kind of distribution to assume for the forecasts. This is closely related to
-        the non-Bayesian losses, e.g.
-        :class:`~.predictive_distributions.MeanFieldNormalPredictiveDistribution`
-        corresponds to MSE loss, while
-        :class:`~.predictive_distributions.CategoricalPredictiveDistribution`
+        the non-Bayesian losses, e.g. :class:`~.distributions.MeanFieldNormal`
+        corresponds to MSE loss, while :class:`~.distributions.Categorical`
         corresponds to cross-entropy loss.
     dataset_size: Optional[int], default: None
         Number of samples in the training dataset. If not provided here it must be
@@ -260,12 +257,15 @@ class AnalyticalKullbackLeiblerLoss(Module):
     :exc:`ValueError`:
         If ``divergence_type`` is ``None`` and the model does not contain any
         :class:`~.VIBaseModule`, i.e. is non-Bayesian.
+    :exc:`UnsupportedDistributionError`:
+        If ``predictive_distribution`` does not support being use as predictive
+        distribution.
     """
 
     def __init__(
         self,
         model: VIModule,
-        predictive_distribution: PredictiveDistribution,
+        predictive_distribution: Distribution,
         dataset_size: Optional[int] = None,
         divergence_type: Optional["KullbackLeiblerModule"] = None,
         heat: float = 1.0,
@@ -276,6 +276,12 @@ class AnalyticalKullbackLeiblerLoss(Module):
         self.dataset_size = dataset_size
         self.heat = heat
         self._track = track
+
+        if not predictive_distribution.is_predictive_distribution:
+            raise UnsupportedDistributionError(
+                f"{predictive_distribution.__class__.__name__} does not support use as"
+                f" predictive distribution"
+            )
 
         if divergence_type is None:
             for module in model.modules():
@@ -294,7 +300,8 @@ class AnalyticalKullbackLeiblerLoss(Module):
                             assert isinstance(kl_type, type(divergence_type))
                         except AssertionError:
                             raise NotImplementedError(
-                                "Handling of inconsistent distributions types is not implemented yet."
[…]
