+++ tests/test_vi/test_priors/__init__.py
@@ -1,1 +1,0 @@
-"""Test torch-bayesian.vi.priors."""
