+++ tests/test_vi/test_variational_distributions/__init__.py
@@ -1,1 +1,0 @@
-"""Test torch-bayesian.vi.variational_distributions."""

[…]
