### 2025-09-22 d7d3060
Update scripts
