+++ tests/test_vi/test_analytical_kl_loss.py
@@ -25,7 +25,7 @@ from torch_bayesian.vi.distributions import (
     NonBayesian,
     UniformPrior,
 )
-from torch_bayesian.vi.utils import use_norm_constants
+from torch_bayesian.vi.utils import UnsupportedDistributionError, use_norm_constants
 
 
 def test_klmodule(device: torch.device) -> None:
@@ -277,10 +277,12 @@ def test_prior_matching(
 
 
 @pytest.mark.parametrize(
-    "prior,var_dist,alt_prior,alt_var_dist,target_kl_module,heat,dataset_size,divergence_type,track,expected_error",
+    "prior,var_dist,predictive_distribution,alt_prior,alt_var_dist,target_kl_module,"
+    "heat,dataset_size,divergence_type,track,expected_error",
     [
         # Test base parameters
         (
+            MeanFieldNormal(),
             MeanFieldNormal(),
             MeanFieldNormal(),
             None,
@@ -293,6 +295,7 @@ def test_prior_matching(
             None,
         ),
         (
+            MeanFieldNormal(),
             MeanFieldNormal(),
             MeanFieldNormal(),
             None,
@@ -307,6 +310,7 @@ def test_prior_matching(
         (
             MeanFieldNormal(),
             MeanFieldNormal(),
+            MeanFieldNormal(),
             None,
             None,
             NormalNormalDivergence,
@@ -317,6 +321,7 @@ def test_prior_matching(
             None,
         ),
         (
+            MeanFieldNormal(),
             MeanFieldNormal(),
             MeanFieldNormal(),
             None,
@@ -332,6 +337,7 @@ def test_prior_matching(
         (
             UniformPrior(),
             MeanFieldNormal(),
+            MeanFieldNormal(),
             None,
             None,
             UniformNormalDivergence,
@@ -344,6 +350,7 @@ def test_prior_matching(
         (
             MeanFieldNormal(),
             NonBayesian(),
+            MeanFieldNormal(),
             None,
             None,
             NonBayesianDivergence,
@@ -356,6 +363,7 @@ def test_prior_matching(
         (
             UniformPrior(),
             NonBayesian(),
+            MeanFieldNormal(),
             None,
             None,
             NonBayesianDivergence,
@@ -369,6 +377,7 @@ def test_prior_matching(
         (
             MeanFieldNormal(),
             MeanFieldNormal(),
+            MeanFieldNormal(),
             UniformPrior(),
             None,
             NormalNormalDivergence,
@@ -379,6 +388,7 @@ def test_prior_matching(
             0,
         ),
         (
+            MeanFieldNormal(),
             MeanFieldNormal(),
             MeanFieldNormal(),
             None,
@@ -393,6 +403,7 @@ def test_prior_matching(
         (
             MeanFieldNormal(),
             MeanFieldNormal(),
+            MeanFieldNormal(),
             UniformPrior(),
             NonBayesian(),
             NormalNormalDivergence,
@@ -405,6 +416,7 @@ def test_prior_matching(
         (
             DummyPrior(),
             MeanFieldNormal(),
+            MeanFieldNormal(),
             None,
             None,
             NormalNormalDivergence,
@@ -417,6 +429,7 @@ def test_prior_matching(
         (
             MeanFieldNormal(),
             DummyVarDist(),
+            MeanFieldNormal(),
             None,
             None,
             NormalNormalDivergence,
@@ -429,6 +442,7 @@ def test_prior_matching(
         (
             DummyPrior(),
             DummyVarDist(),
+            MeanFieldNormal(),
             None,
             None,
             NormalNormalDivergence,
@@ -438,11 +452,24 @@ def test_prior_matching(
             False,
             1,
         ),
-        (None, None, None, None, NormalNormalDivergence, None, None, None, False, 2),
+        (
[…]
