+++ torch_bayesian/vi/distributions/base.py
@@ -13,7 +13,169 @@ if TYPE_CHECKING:
 
 
 class Distribution(metaclass=PostInitCallMeta):
-    """Base class for distributions."""
+    r"""
+    Base class for distributions.
+
+    Distributions for variational implemented as subclasses of this. However, not all
+    distributions can function or make sense in all available roles. Furthermore,
+    each role has distinct requirements, which can be implemented independently.
+    Available roles are prior, variational distribution, and predictive distribution.
+
+    To simplify detection of correct usage this class has the three attributes
+    :attr:`~self.is_prior`, :attr:`~self.is_variational_distribution`, and
+    :attr:`~self.is_predictive_distribution`, which are `False` by default. When
+    implementing custom distributions the flags for the intended usages must be set to
+    `True`. This also implements initialization time checking that provides feedback if
+    any required components for the flagged uses is missing.
+
+    A prior specifies knowledge about the parameter distribution before training. In
+    Bayesian training, weights are generally drawn towards the prior, unless they take
+    an important role. Mathematically, this prior pull can take the same role as weight
+    decay.
+
+
+    Each prior must name the :attr:`~self.distribution_parameters` that define it as
+    well as the way to calculate the log likelihood of a weight configuration in the
+    :meth:`~self.prior_log_prob` method.
+
+    Generally, each name in :attr:`~self.distribution_parameters` should also be an
+    attribute of the class storing that parameter. This is necessary since the prior
+    typically needs to be rescaled based on the layer width. By default, each parameter
+    is assumed to require scaling, but in certain cases like the shape parameter of a
+    Gamma distribution, this might not be the case. It that case the subset of scaling
+    parameters must be specified in :attr:`~self._scaling_parameters`. Non-scaling
+    parameters are technically not required to be a class attribute.
+
+    Furthermore, parameters might only assume positive values. These should be stored as
+    logarithm of their true value, mapping them to the whole real line. Their parameter
+    name should begin with the prefix "log\_", which is automatically detected and
+    handled during rescaling.
+
+    To enable the ``prior_initialization`` functionality, the class must implement the
+    :meth:`~self.reset_parameters` method. Which initializes the parameters for one
+    random variable of a model, whose variational parameters are supported by the prior,
+    to the prior values.
+
+    Parameters
+    ----------
+    distribution_parameters: Tuple[str, ...]
+        Parameters characterizing the prior and can be set during prior based
+        initialization.
+    _required_parameters: Tuple[str, ...], default: ()
+        Parameters besides a sample needed to calculate :meth:`~log_prob`.
+    _scaling_parameters: Tuple[str, ...], default: :attr:`~distribution_parameters`
+        Parameters that need to be rescaled for prior rescaling.
+
+    Methods
+    -------
+    prior_log_prob: Callable[..., Tensor]
+        Function to calculate the log likelihood of a weight configuration under this
+        prior.
+
+
+    A variational distribution specifies the parametrization used to fit the true weight
+    distribution, i.e., the weight posterior.
+
+    Each prior must name the :attr:`~self.distribution_parameters` that will be
+    optimized during training it as well as a default for each parameter in
+    :attr:`~self._default_variational_parameters`. It is important to note that the
+    first parameter is assumed to be a form of mean or mode of the distribution that
+    might be used as the weight in a non-Bayesian version of the network. While a
+    default value for it must be given, it will usually be ignored in initialization in
+    favor of initializing it similar to non-Bayesian weights.
+
+    Additionally, a :meth:`~self.sample` method must be defined that accepts one tensor
+    for each variational parameter and returns a sample from the specified
+    distributions. Finally, the way to calculate the log likelihood of a weight
[…]
