### 2025-09-26 b96ba1b
Plain DDP
