+++ scripts/new_par_sample/DDP_Parallelism/DDP.py
@@ -1,0 +1,156 @@
+import torch.optim as optim
+from torch.nn.parallel import DistributedDataParallel as DDP
+from torch.utils.data import DataLoader, DistributedSampler
+from torchvision import datasets, transforms, models
+import torch
+from torch import Tensor, nn
+import torch_bayesian.vi as vi
+from torch_bayesian.vi.variational_distributions import MeanFieldNormalVarDist
+from torch.utils.data import DataLoader
+from typing import Callable
+import torch.distributed as dist
+import torch.nn.functional as F
+import os
+import torch.multiprocessing as mp
+from torchvision.transforms import ToTensor
+from torchvision import datasets
+import numpy as np
+import random
+sampling_state = None
+
+
+
+def setup(rank, world_size):
+    # Initialize distributed backend
+    dist.init_process_group(
+        backend="nccl",  # Use NCCL for CUDA
+        init_method="env://",
+        rank=rank,
+        world_size=world_size
+    )
+
+def cleanup():
+    """
+    Clean up the process group.
+    """
+    dist.destroy_process_group()
+
+
+def seed_all(seed, rank):
+    seed = seed + rank
+    random.seed(seed)
+    np.random.seed(seed)
+    torch.manual_seed(seed)
+    torch.cuda.manual_seed(seed)
+    torch.cuda.manual_seed_all(seed)
+    torch.backends.cudnn.deterministic = True
+    torch.backends.cudnn.benchmark = False
+
+def DDP_train(train_dataloader, test_dataloader, sampler, model, loss_fn, optimizer, epochs, sample_num, rank, world_size, device):
+    for t in range(epochs):
+        sampler.set_epoch(t) 
+        if rank == 0:
+            print(f"Epoch {t + 1}\n-------------------------------")
+
+        sampler.set_epoch(t)    
+        model.train()
+
+        for batch, (x, y) in enumerate(train_dataloader):
+            # Load batch
+            x, y = x.to(device), y.to(device)
+            optimizer.zero_grad()
+
+
+            # Get predictions
+            pred = model(x, samples=sample_num)
+
+            # Calculate loss
+            mean_model_output = pred.mean(dim=0)
+            loss = loss_fn(mean_model_output, y)
+            # print(loss.item())
+
+            # Backpropagation
+            loss.backward()
+
+            # Update Model
+            optimizer.step()
+        
+    DDP_test(test_dataloader, model, loss_fn, sample_num,rank, world_size, device)
+
+
+    return model
+
+
+def DDP_test(dataloader: DataLoader,
+         model: vi.VIModule,
+         loss_fn: Callable,
+         sample_num,
+         rank,
+         world_size,
+         device
+         ):
+
+    num_batches = len(dataloader)
+    model.eval()
+    test_loss = 0.0
+    correct = 0.0
+    global sampling_state  # Randomness switch
+
+    with torch.no_grad():
+        for x, y in dataloader:
+
+            x, y = x.to(device), y.to(device)
+
+            samples = model(x, samples=sample_num)
+
+            mean_model_output = samples.mean(dim=0)
+            correct += (mean_model_output.argmax(1) == y).type(torch.float).sum().item()
+            test_loss += loss_fn(mean_model_output, y).item()
+
+    if rank == 0:
+        test_loss /= num_batches
+        correct /= len(dataloader.dataset)
+
+        print(
+            f"Test Error: \n Accuracy: {(100 * correct):>0.1f}%, Avg loss: {test_loss:>8f} \n"
+        )
+
+    return
+   
+
[…]
