### 2025-09-29 bab4855
3 Types of Parallelism
