### 2025-10-14 7388924
Distribution overhaul (#32)
Merge VariationalDistribution, Prior, and PredictiveDistribution into a single class. All distribution classes lose the associated suffixes unless they only function as one of the three.
