+++ torch_bayesian/vi/predictive_distributions/__init__.py
@@ -1,13 +1,0 @@
-"""Offers a collection of predictive distributions."""
-
-from .base import PredictiveDistribution
-from .categorical import CategoricalPredictiveDistribution
-from .non_bayesian import NonBayesianPredictiveDistribution
-from .normal import MeanFieldNormalPredictiveDistribution
-
-__all__ = [
-    "PredictiveDistribution",
-    "MeanFieldNormalPredictiveDistribution",
-    "CategoricalPredictiveDistribution",
-    "NonBayesianPredictiveDistribution",
-]
