+++ pyproject.toml
@@ -12,7 +12,7 @@ packages = [
 
 [project]
 name = "torch_bayesian"
-version = "1.0.0"
+version = "0.9.0"
 authors = [
   { name="RAI", email="rai@kit.edu"},
 ]
