### 2025-10-28 42d0252
Tweak stting files, add readthedocs settings
