+++ .github/workflows/release.yaml
@@ -37,7 +37,6 @@ jobs:
           python -m build
       #- name: Publish package distributions to PyPI
       #  # Only run if the tag is a release
-      #  if: github.event_name == 'push' && startsWith(github.ref, 'refs/tags/v') && !contains(github.ref, 'rc')
       #  uses: pypa/gh-action-pypi-publish@76f52bc884231f62b9a034ebfe128415bbaabdfc # release/v1
       - name: Publish package to Test PyPI
         # Only run if the tag is a release candidate
