+++ pyproject.toml
@@ -20,6 +20,7 @@ description = "Library for BNNs"
 readme = "README.md"
 requires-python = ">=3.9"
 classifiers = [
+    "Programming Language :: Python",
     "Programming Language :: Python :: 3",
     "Development Status :: 4 - Beta",
 ]
