### 2025-10-28 721738e
Release preparations
* Fix README errors and add contribution guidelines

* Documentation tweaks

* Add .readthedocs.yaml
