+++ docs/source/conf.py
@@ -16,7 +16,7 @@ sys.path.insert(0, os.path.abspath("../../torch_bayesian"))
 project = "torch_bayesian"
 copyright = "2025, RAI-SCC"
 author = "RAI-SCC"
-release = "0.1"
+release = "0.9.0"
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
