+++ README.md
@@ -2,7 +2,7 @@
 
 This package provides a simple way for non-expert users to implement and train Bayesian
 Neural Networks (BNNs) with Variational Inference (VI). To make this as easy as possible
-most components mirror components from [pytorch](https://pytorch.org/docs/stable/index.html).
+most components mirror components from [PyTorch](https://pytorch.org/docs/stable/index.html).
 
 - [Installation](#installation)
 - [Documentation](#documentation)
@@ -71,9 +71,9 @@ You can then open ``torch_bayesian/docs/build/index.html`` to start browsing.
 
 ## Quickstart
 
-This Quickstart guide assumes basic familiarity with [pytorch](https://pytorch.org/docs/stable/index.html)
+This Quickstart guide assumes basic familiarity with [PyTorch](https://pytorch.org/docs/stable/index.html)
 and knowledge of how to implement the intended model in it. For a (potentially familiar)
-example see `scripts/pytorch_tutorial.py`, which contains a copy of the pytorch
+example see `scripts/pytorch_tutorial.py`, which contains a copy of the PyTorch
 [Quickstart tutorial](https://pytorch.org/tutorials/beginner/basics/quickstart_tutorial.html)
 modified to train a BNN with variational inference.
 Three levels are introduced:
@@ -86,13 +86,13 @@ Three levels are introduced:
 
 Many parts of a neural network remain completely unchanged when turning it into a BNN.
 Indeed, only `Module`s containing `nn.Parameter`s, need to be changed. Therefore, if a
-pytorch model fulfills two requirements it can be transferred almost unchanged:
+PyTorch model fulfills two requirements it can be transferred almost unchanged:
 
-1. All pytorch `Module`s containing parameters have equivalents in this package (table below).
+1. All PyTorch `Module`s containing parameters have equivalents in this package (table below).
 2. The model can be expressed purely as a sequential application of a list of layers,
 i.e. with `nn.Sequential`.
 
-| pytorch          | vi replacement  |
+| PyTorch          | vi replacement  |
 |------------------|-----------------|
 | `nn.Linear`      | `VILinear`      |
 | `nn.Conv1d`      | `VIConv1d`      |
@@ -103,7 +103,7 @@ i.e. with `nn.Sequential`.
 Given these two conditions, inherit the module from `vi.VIModule` instead of `nn.Module`
 and use `vi.VISequential` instead of `nn.Sequential`. Then replace all layers
 containing parameters as shown in the table above. For basic usage initialize these
-modules with the same arguments as their pytorch equivalent. For advanced usage see
+modules with the same arguments as their PyTorch equivalent. For advanced usage see
 [Quickstart: Level 2](#level-2). Many other layers can be included as-is. In particular
 activation functions, pooling, and padding (even dropout, though they
 should not be necessary since the prior acts as regularization). Currently not supported
@@ -111,20 +111,18 @@ are recurrent and transposed convolution layers. Normalization layers may
 have parameters depending on their setting, but can likely be left non-Bayesian.
 
 Additionally, the loss must be replaced. To start out use `vi.KullbackLeiblerLoss`,
-which requires a `PredictiveDistribution` and the size of the training dataset (this is
-important for balancing of assumptions and data, more details
-[here](#variational-inference)). Choose your `PredictiveDistribution`
-from the table below based on the loss you would use in pytorch (more details
-[here](#the-predictive-distribution)).
+which requires a `Distribution` with `self.is_predictive_distribution=True` and the size
+of the training dataset (this is important for balancing of assumptions and data. Choose
+your `Distribution` from the table below based on the loss you would use in PyTorch.
 
 > [!IMPORTANT]
 > `KullbackLeiblerLoss` requires the length of the dataset, not the dataloader, which is
[…]
