+++ pyproject.toml
@@ -12,7 +12,7 @@ packages = [
 
 [project]
 name = "torch_bayesian"
-version = "1.0.0"
+version = "0.9.0"
 authors = [
   { name="RAI", email="rai@kit.edu"},
 ]
@@ -20,6 +20,7 @@ description = "Library for BNNs"
 readme = "README.md"
 requires-python = ">=3.9"
 classifiers = [
+    "Programming Language :: Python",
     "Programming Language :: Python :: 3",
     "Development Status :: 4 - Beta",
 ]
