+++ paper/paper.bib
@@ -1,104 +1,0 @@
-@article{bingham2019pyro,
-  author    = {Eli Bingham and
-               Jonathan P. Chen and
-               Martin Jankowiak and
-               Fritz Obermeyer and
-               Neeraj Pradhan and
-               Theofanis Karaletsos and
-               Rohit Singh and
-               Paul A. Szerlip and
-               Paul Horsfall and
-               Noah D. Goodman},
-  title     = {Pyro: Deep Universal Probabilistic Programming},
-  journal   = {J. Mach. Learn. Res.},
-  volume    = {20},
-  pages     = {28:1--28:6},
-  year      = {2019},
-  url       = {http://jmlr.org/papers/v20/18-403.html}
-}
-
-@article{stan2025stan,
-  author    = {{Stan Development Team}},
-  title     = {Stan Reference Manual, 2.37},
-  year      = {2025},
-  url       = {https://mc-stan.org},
-}
-
-@inproceedings{ansel2024pytorch,
-  author = {Ansel, Jason and Yang, Edward and He, Horace and Gimelshein, Natalia and Jain, Animesh and Voznesensky, Michael and Bao, Bin and Bell, Peter and Berard, David and Burovski, Evgeni and Chauhan, Geeta and Chourdia, Anjali and Constable, Will and Desmaison, Alban and DeVito, Zachary and Ellison, Elias and Feng, Will and Gong, Jiong and Gschwind, Michael and Hirsh, Brian and Huang, Sherlock and Kalambarkar, Kshiteej and Kirsch, Laurent and Lazos, Michael and Lezcano, Mario and Liang, Yanbo and Liang, Jason and Lu, Yinghai and Luk, C. K. and Maher, Bert and Pan, Yunjie and Puhrsch, Christian and Reso, Matthias and Saroufim, Mark and Siraichi, Marcos Yukio and Suk, Helen and Zhang, Shunting and Suo, Michael and Tillet, Phil and Zhao, Xu and Wang, Eikan and Zhou, Keren and Zou, Richard and Wang, Xiaodong and Mathews, Ajit and Wen, William and Chanan, Gregory and Wu, Peng and Chintala, Soumith},
-  title = {PyTorch 2: Faster Machine Learning Through Dynamic Python Bytecode Transformation and Graph Compilation},
-  year = {2024},
-  isbn = {9798400703850},
-  publisher = {Association for Computing Machinery},
-  address = {New York, NY, USA},
-  url = {https://doi.org/10.1145/3620665.3640366},
-  doi = {10.1145/3620665.3640366},
-  abstract = {This paper introduces two extensions to the popular PyTorch machine learning framework, TorchDynamo and TorchInductor, which implement the torch.compile feature released in PyTorch 2. TorchDynamo is a Python-level just-in-time (JIT) compiler that enables graph compilation in PyTorch programs without sacrificing the flexibility of Python. It achieves this by dynamically modifying Python bytecode before execution and extracting sequences of PyTorch operations into an FX graph, which is then JIT compiled using one of many extensible backends. TorchInductor is the default compiler backend for TorchDynamo, which translates PyTorch programs into OpenAI's Triton for GPUs and C++ for CPUs. Results show that TorchDynamo is able to capture graphs more robustly than prior approaches while adding minimal overhead, and TorchInductor is able to provide a 2.27\texttimes{} inference and 1.41\texttimes{} training geometric mean speedup on an NVIDIA A100 GPU across 180+ real-world models, which outperforms six other compilers. These extensions provide a new way to apply optimizations through compilers in eager mode frameworks like PyTorch.},
-  booktitle = {Proceedings of the 29th ACM International Conference on Architectural Support for Programming Languages and Operating Systems, Volume 2},
-  pages = {929–947},
-  numpages = {19},
[…]
