### 2025-10-28 661b499
Beta 0.9.0 Test release
* Merge VIModule and VIBaseModule into one class

* Add internal log_prob handling. log prob handling is now completely hidden from the user, simplifying the interface

* Distribution overhaul, all distributions now use the same base class

* Set up github workflows, contribution guidelines, and readthedocs settings

* Fix fix and update various typos or changes in docs
