+++ torch_bayesian/vi/base.py
@@ -1,10 +1,11 @@
 import warnings
 from copy import deepcopy
-from typing import Any, Callable, Dict, List, Optional, Tuple
+from typing import Any, Callable, Dict, List, Optional, Tuple, Union, cast
 
 import torch
 import torch.utils.hooks as hooks
 from torch import Tensor
+from torch._C._functorch import get_unwrapped
 from torch.nn import Module, Parameter, init
 from torch.nn.common_types import _tensor_list_t
 from torch.nn.modules.module import (
@@ -16,11 +17,13 @@ from torch.nn.modules.module import (
     _WrappedHook,
 )
 
-from .utils import PostInitCallMeta
-from .utils.common_types import VIReturn, _prior_any_t, _vardist_any_t
+from .distributions import MeanFieldNormal
+from .utils import NoVariablesError, PostInitCallMeta, UnsupportedDistributionError
+from .utils.common_types import _dist_any_t
+from .utils.vi_return import VIReturn
 
 
-def _forward_unimplemented(self: Module, *input_: Optional[Tensor]) -> VIReturn[Tensor]:
+def _forward_unimplemented(self: Module, *input_: Optional[Tensor]) -> Tensor:
     r"""Define the computation performed at every call.
 
     Should be overridden by all subclasses.
@@ -41,19 +44,31 @@ class VIModule(Module, metaclass=PostInitCallMeta):
     """
     Base class for Modules using Variational Inference.
 
-    Conceptually, this class takes the place of :class:`torch.nn.Module` for BNNs. It is
-    used for any module that has no Bayesian parameters of its own. If the module should
-    have Bayesian parameters, use :class:`~.VIBaseModule` instead, which is an extension
-    of this class.
+    This class takes the place of :class:`torch.nn.Module` for BNNs. It is used for any
+    Bayesian module. While a :class:`torch.nn.Module` may be contained within a
+    :class:`~.VIModule` and vice versa there should always be a singular
+    :class:`~.VIModule` containing all others to avoid superfluous sampling dimensions.
 
     :class:`~.VIModule` contains some additional functionality. Firstly, it keeps track
     of whether the model should return the log probability of the sampled weight (i.e.,
-    the log likelihood to obtain these specific values when sampling from the prior or
+    the log probability to obtain these specific values when sampling from the prior or
     variational distribution).These are needed for loss calculation by
-    :class:`~.KullbackLeiblerLoss`. If your module contains multiple submodules make
-    sure to add all log probabilities and return them, if required by
-    :attr:`~self._return_log_probs`. This attribute can be changed with
-    :meth:`~return_log_probs`.
+    :class:`~.KullbackLeiblerLoss`. This is handled automatically as part of accessing
+    the weight matrices, if :attr:`~self.return_log_probs` is True. For evaluation, it
+    can be set to `False`.
+
+    The outermost module will aggregate the log probabilities and pack the output and
+    log probs into a :class:`~.VIReturn` object, which behaves like a pytorch Tensor.
+    However, it has the added attribute :attr:`log_probs` where the log probs are
+    stored. This is used by losses, therefore it is easiest to wrap all operations into
+    a :class:`~.VIModule` and feed the output directly into a loss function. Classical
+    losses will treat it like a Tensor and `torch_bayesian` losses can use the log prob
+    information. If the model has multiple output Tensors, each will contain the full
+    log prob information.
+
+    .. IMPORTANT:: When defining custom modules with weights make sure to retrieve them
+        using :meth:`~self.sample_variables` as this will maintain the automatic log
+        prob tracking.
 
     Secondly, a model of nested :class:`~.VIModule` automatically identifies the
[…]
