### 2025-10-31 c435f6c
Refactor from torch_bayesian to TorchBuq
