+++ tests/test_vi/test_distributions/test_distributions_base.py
@@ -6,7 +6,7 @@ import torch
 from torch import Tensor
 from torch.nn import Linear, Module, Parameter
 
-from torch_bayesian.vi.distributions import Distribution
+from torchbuq.vi.distributions import Distribution
 
 
 class TestDistribution:
