+++ docs/source/conf.py
@@ -7,13 +7,13 @@
 import os
 import sys
 
-sys.path.insert(0, os.path.abspath("../../torch_bayesian"))
+sys.path.insert(0, os.path.abspath("../../torchbuq"))
 
 
 # -- Project information -----------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
-project = "torch_bayesian"
+project = "torchbuq"
 copyright = "2025, RAI-SCC"
 author = "RAI-SCC"
 release = "0.9.0"
@@ -30,14 +30,14 @@ autodoc_type_aliases = {
     "torch.nn.common_types._size_1_t": "int | Tuple[int,]",
     "torch.nn.common_types._size_2_t": "int | Tuple[int, int]",
     "torch.nn.common_types._size_3_t": "int | Tuple[int, int, int]",
-    "torch_bayesian.vi.utils.common_types.VIReturn[torch.Tensor]": "torch.Tensor | Tuple[torch.Tensor, torch.Tensor]",
-    "torch_bayesian.vi.utils.common_types.VIReturn[Tuple[torch.Tensor, Optional[torch.Tensor]]]": "Tuple[torch.Tensor, Optional[torch.Tensor]] | Tuple[Tuple[torch.Tensor, torch.Tensor | None], torch.Tensor]",
-    "torch_bayesian.vi.utils.common_types.VIReturn[torch.nn.common_types._tensor_list_t]": "Union[torch.Tensor, List[torch.Tensor]] | Tuple[Union[torch.Tensor, List[torch.Tensor]], torch.Tensor]",
-    "torch_bayesian.vi.utils.common_types._vardist_any_t": "VariationalDistribution | List[VariationalDistribution]",
-    "torch_bayesian.vi.utils.common_types._prior_any_t": "Prior | List[Prior]",
-    "torch_bayesian.vi.utils.common_types._log_prob_return_format[torch.Tensor]": "Tuple[torch.Tensor, torch.Tensor]",
+    "torchbuq.vi.utils.common_types.VIReturn[torch.Tensor]": "torch.Tensor | Tuple[torch.Tensor, torch.Tensor]",
+    "torchbuq.vi.utils.common_types.VIReturn[Tuple[torch.Tensor, Optional[torch.Tensor]]]": "Tuple[torch.Tensor, Optional[torch.Tensor]] | Tuple[Tuple[torch.Tensor, torch.Tensor | None], torch.Tensor]",
+    "torchbuq.vi.utils.common_types.VIReturn[torch.nn.common_types._tensor_list_t]": "Union[torch.Tensor, List[torch.Tensor]] | Tuple[Union[torch.Tensor, List[torch.Tensor]], torch.Tensor]",
+    "torchbuq.vi.utils.common_types._vardist_any_t": "VariationalDistribution | List[VariationalDistribution]",
+    "torchbuq.vi.utils.common_types._prior_any_t": "Prior | List[Prior]",
+    "torchbuq.vi.utils.common_types._log_prob_return_format[torch.Tensor]": "Tuple[torch.Tensor, torch.Tensor]",
     "Ellipsis": "...",
-    "torch_bayesian.vi.base.": "torch_bayesian.vi.",
+    "torchbuq.vi.base.": "torchbuq.vi.",
     #    "torch.Tensor": "Tensor",
 }
 
@@ -57,7 +57,7 @@ extensions = [
 ]
 
 autoapi_type = "python"
-autoapi_dirs = ["../../torch_bayesian/"]
+autoapi_dirs = ["../../torchbuq/"]
 autoapi_ignore = ["*/tests/*"]  # , "*[!_].py"]
 autoapi_template_dir = "_templates/"
[…]
