### 2025-10-31 6e849a4
Further README tweaks
