### 2025-10-31 cac4042
Add revised JOSS submission
