### 2025-10-31 2a79242
Add permissions to paper workflow
