+++ .github/workflows/draft-pdf.yaml
@@ -7,6 +7,8 @@ on:
 
 jobs:
   paper:
+    permissions:
+      contents: read
     runs-on: ubuntu-latest
     name: Paper Draft
     steps:
