### 2025-10-31 156c56a
Second refactor
