+++ torch_blue/vi/utils/vi_return.py
@@ -11,9 +11,9 @@ class VIReturn(Tensor):
     purposes, but provides the optional attribute :attr:`~log_probs`. However, it should
     not be used as replacement since certain pytorch operation will lose the log prob
     information. It is almost exclusively used as the return formate for
-    :class:`~torchbuq.vi.VIModule`s. This allows the output to be treated like a
+    :class:`~torch_blue.vi.VIModule`s. This allows the output to be treated like a
     :class:`torch.Tensor`, but still provide the log prob information when needed.
-    `torchbuq` losses may require this format as input.
+    `torch_blue` losses may require this format as input.
     """
 
     log_probs: Optional[Tensor] = None
