+++ tests/test_vi/test_distributions/test_quiet.py
@@ -4,8 +4,8 @@ import torch
 from pytest import mark
 from torch.nn import Module, Parameter
 
-from torchbuq.vi.distributions import BasicQuietPrior
-from torchbuq.vi.utils import use_norm_constants
+from torch_blue.vi.distributions import BasicQuietPrior
+from torch_blue.vi.utils import use_norm_constants
 
 
 @mark.parametrize("norm_constants", [True, False])
