+++ tests/test_vi/test_distributions/__init__.py
@@ -1,1 +1,1 @@
-"""Test torch-bayesian.vi.distributions."""
+"""Test torch_blue.vi.distributions."""
