+++ tests/test_vi/test_conv.py
@@ -1,8 +1,8 @@
 import torch
 
-from torchbuq.vi import VIConv1d, VIConv2d, VIConv3d
-from torchbuq.vi.conv import _VIConvNd
-from torchbuq.vi.distributions import MeanFieldNormal
+from torch_blue.vi import VIConv1d, VIConv2d, VIConv3d
+from torch_blue.vi.conv import _VIConvNd
+from torch_blue.vi.distributions import MeanFieldNormal
 
 
 # Since these are basically copied from torch 2.4 testing is limited
