+++ torch_blue/vi/distributions/quiet.py
@@ -109,7 +109,7 @@ class BasicQuietPrior(Distribution):
         variable: str
             The name of the random variable to reset as given by
             :attr:`variational_parameters` of the associated
-            :class:`~torch_blue.vi.variational_distributions.VariationalDistribution`.
+            :class:`~torch_blue.vi.distributions.Distribution`.
 
         Returns
         -------
