### 2025-10-31 00b2bd2
Main refactor (#37)
* Refactor from torch_bayesian to torch_blue
