### 2025-11-03 1a7b1ec
Refactor dev branch and paper to new name (#40)
* Refactor from torch_bayesian to torch_blue

* Re-add paper
