+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'torch_bayesian: A Flexible Python Package for Bayesian Neural Networks in PyTorch'
+title: 'torch_blue: A Flexible Python Package for Bayesian Neural Networks in PyTorch'
 tags:
     - Python
     - Bayesian Neural Networks
@@ -27,7 +27,7 @@ affiliations:
       ror: 04t3en479
     - name: Helmholtz AI
       index: 2
-date: 08 September 2025
+date: 31 October 2025
 bibliography: paper.bib
 ---
 
@@ -47,7 +47,7 @@ directly supports direct BNN model prototyping by offering pre-programmed Bayesi
 network layer types, similar to PyTorch's `nn` module. This forces any BNNs to be
 implemented from scratch, which can be challenging even for non-Bayesian networks.
 
-`torch_bayesian` addresses this by providing an interface that is almost identical to
+`torch_blue` addresses this by providing an interface that is almost identical to
 the widely used PyTorch [@ansel2024pytorch] for basic use, providing a low entry
 barrier, as well as an advanced interface designed for exploration and research.
 Overall, this allows users to set up models and even custom layers without worrying
@@ -57,7 +57,7 @@ about the Bayesian intricacies under the hood.
 
 To represent uncertainty BNNs do not consider their weights as point values, but
 random variables, i.e., distributions. The optimization goal becomes adapting the weight
-distributions to minimize their distance to the perfect distribution. This requires two
+distributions to minimize their distance to the true distribution. This requires two
 assumptions. For one, the distance between distributions needs to be defined, for
 which the Kullback-Leibler divergence [@kullback1951information] is typically used.
 Secondly, optimizing an object as complex as a distribution is a non-trivial task.
@@ -77,13 +77,13 @@ concepts - like plate notation - which will be unfamiliar to many primary machin
 learning users.
 
 ![Code example of a three-layer Bayesian MLP with cross-entropy loss in
-`torch_bayesian`. The highlight colors relate user-facing components to their position
+`torch_blue`. The highlight colors relate user-facing components to their position
 in \autoref{design_graph}. \label{code}](code_example.png)
 
-![Design graph of `torch_bayesian` colored highlights correspond to their practical
+![Design graph of `torch_blue` colored highlights correspond to their practical
 applications in the code example (\autoref{code}). \label{design_graph}](design_graph.png)
 
-`torch_bayesian` sacrifices this extreme flexibility to allow nearly fully automating
+`torch_blue` sacrifices this extreme flexibility to allow nearly fully automating
 VI with reparametrization (Bayes by Backprop) [@blundell15bbb]. The ability to use
 multiple independent sampling dimensions is removed, which allows to fully automate a
 single sampling dimensions in the outermost instance of the new base class `VIModule`,
@@ -94,18 +94,18 @@ whenever weights are sampled, aggregated, and returned once again by the outermo
 
 # Core design and features
 
-`torch_bayesian` is designed around two core aims:
+`torch_blue` is designed around two core aims:
 
 1. Ease of use, even for users with little to no experience with Bayesian statistics
 2. Flexibility and extensibility as required for research and exploration
 
-While ease of use colors all design decisions, it features most prominently in the
+While ease of use influences all design decisions, it features most prominently in the
 PyTorch-like interface. While currently only the most common layer types provided by
 PyTorch are supported, corresponding Bayesian layers follow an analogous naming
 pattern and accept the same arguments as their PyTorch version. Additionally, while
 there are minor differences the process of implementing custom layers is also very
 similar to PyTorch. To illustrate this \autoref{code} and \ref{design_graph} show an
-application example and internal interactions of `torch_bayesian` with the colors
+application example and internal interactions of `torch_blue` with the colors
 connecting the abstract and applied components.
 
[…]
