+++ tests/test_vi/test_utils/test_init.py
@@ -1,7 +1,7 @@
 import torch
 from torch.nn import Module, Parameter
 
-from torch_bayesian.vi.utils import init
+from torch_blue.vi.utils import init
 
 
 def test_fixed(device: torch.device) -> None:
