+++ tests/test_vi/test_utils/test_use_norm_constants.py
@@ -1,5 +1,5 @@
-from torch_bayesian.vi import _globals
-from torch_bayesian.vi.utils import use_norm_constants
+from torch_blue.vi import _globals
+from torch_blue.vi.utils import use_norm_constants
 
 
 def test_use_norm_constants() -> None:
