### 2025-11-03 #8968 @ArvidWeyrauch
Hello @logological - thank you for your time as editor.
There have been two cosmetic changes in the process of the PyPI release. I hope these won't cause issues.
Firstly, we decided to consider this version 0.9.0 until we had external user feedback.
Secondly, the package name was claimed by someone else on PyPI about two weeks ago. Therefore we refactored to the name torch_blue. I have adjusted the paper on the submitted dev branch to reflect this.

If you need additional suggestions for reviewers: elcorto and JSchmie have shown interest in this package, when I introduce an earlier version at a conference as poster, but are otherwise unrelated to our group.
