### 2025-11-03 #8968 @logological
@ArvidWeyrauch Thanks for the update.  It's no problem to continue to make changes to the software and paper as long as it's not too much of a moving target for the reviewers and myself.  I'll now begin the process of recruiting reviewers, so thank you for your nominations.  Once we've got two or three confirmed, I'll open a separate issue for the paper review.
