### 2025-11-04 f30b07d
Add basic autoconversion
