+++ torch_blue/vi/utils/convert.py
@@ -1,0 +1,91 @@
+from typing import Dict, Optional, Tuple, cast
+
+import torch
+from torch import nn
+
+from ..base import VIModule
+from ..distributions import MeanFieldNormal
+from .common_types import VIkwargs, _dist_any_t
+
+_blacklist = [nn.ReLU, nn.LayerNorm]
+
+
+def _convert_module(
+    module: nn.Module,
+    variational_distribution: _dist_any_t = MeanFieldNormal(),
+    prior: _dist_any_t = MeanFieldNormal(),
+    rescale_prior: bool = False,
+    kaiming_initialization: bool = True,
+    prior_initialization: bool = False,
+    return_log_probs: bool = True,
+    device: Optional[torch.device] = None,
+    dtype: Optional[torch.dtype] = None,
+) -> None:
+    if module.__class__ in _blacklist:
+        return
+
+    vikwargs: VIkwargs = dict(
+        variational_distribution=variational_distribution,
+        prior=prior,
+        rescale_prior=rescale_prior,
+        kaiming_initialization=kaiming_initialization,
+        prior_initialization=prior_initialization,
+        return_log_probs=return_log_probs,
+        device=device,
+        dtype=dtype,
+    )
+    new_class_name = "AVI" + module.__class__.__name__
+    new_class = type(new_class_name, (VIModule, module.__class__), dict())
+    setattr(new_class, "forward", module.forward)
+
+    module.__class__ = new_class
+    module = cast(VIModule, module)
+    if len(module._parameters) == 0:
+        VIModule.__post_init__(module)
+        return
+
+    # TODO: Add weight copying option
+    # primary_parameter = variational_distribution.primary_parameter
+    parameters = module._parameters
+    module._parameters = dict()
+
+    variable_shapes: Dict[str, Optional[Tuple[int, ...]]] = dict()
+    for name, parameter in parameters.items():
+        if parameter is None:
+            variable_shapes[name] = None
+        else:
+            variable_shapes[name] = tuple(parameter.shape)
+
+    VIModule.__init__(module, variable_shapes, convert_overwrite=True, **vikwargs)
+    VIModule.__post_init__(module)
+
+
+def convert_to_vimodule(
+    module: nn.Module,
+    variational_distribution: _dist_any_t = MeanFieldNormal(),
+    prior: _dist_any_t = MeanFieldNormal(),
+    rescale_prior: bool = False,
+    kaiming_initialization: bool = True,
+    prior_initialization: bool = False,
+    return_log_probs: bool = True,
+    device: Optional[torch.device] = None,
+    dtype: Optional[torch.dtype] = None,
+) -> None:
+    """Convert a PyTorch module to a VIModule."""
+    vikwargs: VIkwargs = dict(
+        variational_distribution=variational_distribution,
+        prior=prior,
+        rescale_prior=rescale_prior,
+        kaiming_initialization=kaiming_initialization,
+        prior_initialization=prior_initialization,
+        return_log_probs=return_log_probs,
+        device=device,
+        dtype=dtype,
+    )
+    for m in module.modules():
+        _convert_module(m, **vikwargs)
+
+    # __post_init__ needs to be called here again since the loop above goes outside in
+    # and __post_init__ relies on being called the last time by the outermost module
+    module = cast(VIModule, module)
+    VIModule.__post_init__(module)
