### 2025-11-04 9da88bb
Add basic autoconversion
