### 2025-11-04 04f240a
Add alpha weight keeping conversion
