### 2025-11-04 be49997
Add weight keeping test and debug
