+++ docs/source/installation.rst
@@ -1,0 +1,9 @@
+.. _installation:
+
+Installation
+============
+
+.. include:: ../../README.md
+   :parser: myst
+   :start-after: ## Installation
+   :end-before: ## Documentation
