### 2025-11-05 012e07e
Implement forward rerouting
