+++ torch_blue/vi/base.py
@@ -1,21 +1,11 @@
-import warnings
 from copy import deepcopy
 from typing import Any, Callable, Dict, List, Optional, Tuple, Union, cast
 
 import torch
-import torch.utils.hooks as hooks
 from torch import Tensor
 from torch._C._functorch import get_unwrapped
 from torch.nn import Module, Parameter, init
 from torch.nn.common_types import _tensor_list_t
-from torch.nn.modules.module import (
-    _global_backward_hooks,
-    _global_backward_pre_hooks,
-    _global_forward_hooks,
-    _global_forward_hooks_always_called,
-    _global_forward_pre_hooks,
-    _WrappedHook,
-)
 
 from .distributions import MeanFieldNormal
 from .utils import NoVariablesError, PostInitCallMeta, UnsupportedDistributionError
@@ -433,7 +423,7 @@ class VIModule(Module, metaclass=PostInitCallMeta):
             One or multiple Tensors with log prob annotation
         """
         expanded = [self._expand_to_samples(x, samples=samples) for x in input_]
-        out: _tensor_list_t = torch.vmap(self.forward, randomness="different")(
+        out: _tensor_list_t = torch.vmap(self._module_forward, randomness="different")(
             *expanded, **kwargs
         )
 
@@ -509,6 +499,12 @@ class VIModule(Module, metaclass=PostInitCallMeta):
                 module._has_sampling_responsibility = False
         self._has_sampling_responsibility = True
 
+    def _reroute_forward(
+        self, new_forward: Callable, new_name: str = "_module_forward"
+    ) -> None:
+        setattr(self, new_name, self.forward)
+        self.forward = new_forward
+
     def __post_init__(self) -> None:
         """
         After __init__ set sampling responsibility.
@@ -518,196 +514,14 @@ class VIModule(Module, metaclass=PostInitCallMeta):
         by default.
         """
         self._set_sampling_responsibility()
+        self._reroute_forward(self._pre_forward)
 
     def _pre_forward(self, *input_: Any, **kwargs: Any) -> Any:
         """Select sampled_forward or forward based on _has_sampling_responsibility."""
         if self._has_sampling_responsibility:
             return self.sampled_forward(*input_, **kwargs)
         else:
-            return self.forward(*input_, **kwargs)
-
-    # Copied from pytorch 2.4, basically untested since assumed working
-    # Change: call _pre_forward instead of forward
-    def _slow_forward(self, *input_: Any, **kwargs: Any) -> Any:  # pragma: no cover
-        tracing_state = torch._C._get_tracing_state()
-        forward_call = self._pre_forward
-        if not tracing_state or isinstance(forward_call, torch._C.ScriptMethod):
-            return forward_call(*input_, **kwargs)
-        recording_scopes = torch.jit._trace._trace_module_map is not None
-        if recording_scopes:
-            # type ignore was added because at this point one knows that
-            # torch.jit._trace._trace_module_map is not Optional and has type Dict[Any, Any]
-            name = (
-                torch.jit._trace._trace_module_map[self]
-                if self in torch.jit._trace._trace_module_map
-                else None
-            )  # type: ignore[index, operator] # noqa: B950
-            if name:
-                tracing_state.push_scope(name)
-            else:
-                recording_scopes = False
-        try:
-            result = forward_call(*input_, **kwargs)
-        finally:
-            if recording_scopes:
-                tracing_state.pop_scope()
-        return result
-
-    # Copied from pytorch 2.4, basically untested since assumed working
-    # Change: call _pre_forward instead of forward
-    def _call_impl(self, *args: Any, **kwargs: Any) -> Any:  # pragma: no cover
-        if torch._C._get_tracing_state():
-            forward_call = self._slow_forward
-        else:
[…]
