### 2025-11-05 9a5d77e
Patch AnalyticalKLLoss to work with empty random variables
