+++ torch_blue/vi/base.py
@@ -1,6 +1,6 @@
 import warnings
 from copy import deepcopy
-from typing import Any, Callable, Dict, List, Optional, Tuple, Union, cast
+from typing import Any, Callable, Dict, List, Mapping, Optional, Tuple, Union, cast
 
 import torch
 import torch.utils.hooks as hooks
@@ -111,7 +111,7 @@ class VIModule(Module, metaclass=PostInitCallMeta):
 
     Parameters
     ----------
-    variable_shapes: Optional[Dict[str, Optional[Tuple[int, ...]]]], default = None
+    variable_shapes: Optional[Mapping[str, Optional[Tuple[int, ...]]]], default = None
         Shape specifications for all random variables. Keys are turned into
         :attr:`self.random_variables` in insertion order.
     VIkwargs
@@ -134,7 +134,7 @@ class VIModule(Module, metaclass=PostInitCallMeta):
 
     def __init__(
         self,
-        variable_shapes: Optional[Dict[str, Optional[Tuple[int, ...]]]] = None,
+        variable_shapes: Optional[Mapping[str, Optional[Tuple[int, ...]]]] = None,
         variational_distribution: _dist_any_t = MeanFieldNormal(),
         prior: _dist_any_t = MeanFieldNormal(),
         rescale_prior: bool = False,
@@ -224,7 +224,7 @@ class VIModule(Module, metaclass=PostInitCallMeta):
         return tuple(self.variational_distribution.keys())
 
     def _rescale_prior(
-        self, variable_shapes: Dict[str, Optional[Tuple[int, ...]]]
+        self, variable_shapes: Mapping[str, Optional[Tuple[int, ...]]]
     ) -> None:
         """
         Rescale the prior parameters based on the layer width.
@@ -233,7 +233,7 @@ class VIModule(Module, metaclass=PostInitCallMeta):
 
         Parameters
         ----------
-        variable_shapes: Dict[str, Optional[Tuple[int, ...]]]
+        variable_shapes: Mapping[str, Optional[Tuple[int, ...]]]
             The dictionary of random variable names and shapes as passed to __init__.
 
         Returns
@@ -717,7 +717,7 @@ class VIModule(Module, metaclass=PostInitCallMeta):
         return super().__getattr__(name)
 
     def _calculate_fan_in(
-        self, variable_shapes: Optional[Dict[str, Optional[Tuple[int, ...]]]] = None
+        self, variable_shapes: Optional[Mapping[str, Optional[Tuple[int, ...]]]] = None
     ) -> int:
         if variable_shapes is None:
             loop = self.random_variables
