### 2025-11-06 70c1396
Fill out coverage and upgrade dtype handling of conversion
Add test coverage for None-parameter path of analytical KL-loss.
Make autoconversion copy dtypes and devices from the original
module.
