### 2025-11-06 4ba66b2
Add log_prob reset before forward pass
