### 2025-11-06 62a2fc9
Add log_prob reset before forward pass
