### 2025-11-08 287540d
Implement basic module reuse for conversion
