### 2025-11-08 177c6f6
Implement basic module reuse for conversion
