### 2025-11-08 1f0304d
Implement basic module reuse for conversion
