+++ tests/test_vi/test_transformer.py
@@ -55,244 +55,244 @@ class Filter(VIModule):
             True,
             None,
         ),
-        (
-            32,
-            3,
-            MeanFieldNormal(std=1e-20),
-            3,
-            5,
-            7,
-            False,
-            False,
-            True,
-            False,
-            False,
-            None,
-            None,
-            True,
-            1,
-        ),
-        (
-            32,
-            4,
-            MeanFieldNormal(std=1e-20),
-            3,
-            5,
-            7,
-            False,
-            False,
-            True,
-            False,
-            False,
-            None,
-            None,
-            True,
-            None,
-        ),
-        (
-            32,
-            2,
-            MeanFieldNormal(std=1e-20),
-            3,
-            5,
-            7,
-            True,
-            False,
-            True,
-            False,
-            False,
-            None,
-            None,
-            True,
-            None,
-        ),
-        (
-            32,
-            2,
-            MeanFieldNormal(std=1e-20),
-            3,
-            5,
-            7,
-            True,
-            True,
-            True,
-            False,
-            False,
-            None,
-            None,
-            True,
-            None,
-        ),
-        (
-            32,
-            2,
-            MeanFieldNormal(std=1e-20),
-            3,
-            5,
-            7,
-            False,
-            True,
-            True,
-            False,
-            False,
-            None,
-            None,
-            True,
-            None,
-        ),
-        (
-            32,
-            2,
-            MeanFieldNormal(std=1e-20),
-            3,
-            5,
-            7,
-            False,
-            False,
-            False,
-            False,
-            False,
-            None,
-            None,
-            True,
-            None,
-        ),
-        (
-            32,
-            2,
-            MeanFieldNormal(std=1e-20),
-            3,
-            5,
-            7,
-            False,
-            False,
-            True,
-            True,
-            False,
-            None,
-            None,
-            True,
-            None,
-        ),
-        (
-            32,
-            2,
-            MeanFieldNormal(std=1e-20),
-            3,
-            5,
-            7,
-            False,
-            False,
-            False,
-            True,
-            False,
-            None,
-            None,
-            True,
-            None,
-        ),
-        (
-            32,
-            2,
-            MeanFieldNormal(std=1e-20),
-            3,
-            5,
-            7,
-            False,
-            False,
-            True,
-            False,
-            True,
-            None,
-            None,
-            True,
-            None,
-        ),
-        (
-            32,
-            2,
-            MeanFieldNormal(std=1e-20),
-            3,
-            5,
-            7,
-            False,
-            False,
-            True,
-            False,
-            False,
-            None,
-            None,
-            False,
-            None,
-        ),
-        (
-            32,
-            2,
-            MeanFieldNormal(std=1e-20),
-            3,
-            5,
-            7,
-            False,
-            False,
-            True,
-            False,
-            False,
-            30,
-            None,
-            True,
-            None,
-        ),
-        (
-            32,
-            2,
-            MeanFieldNormal(std=1e-20),
-            3,
-            5,
-            7,
-            False,
[…]
