### 2025-11-09 c79ec16
Add convert class reuse and ban list interces
