+++ tests/test_vi/test_utils/test_convert.py
@@ -1,9 +1,13 @@
+from copy import deepcopy
+
 import pytest
 import torch
 from torch import nn
 
+import torch_blue
 from torch_blue.vi import convert_to_vimodule
-from torch_blue.vi.distributions import NonBayesian
+from torch_blue.vi.distributions import MeanFieldNormal, NonBayesian, StudentT
+from torch_blue.vi.utils import convert
 
 
 @pytest.mark.parametrize(
@@ -21,9 +25,39 @@ def test_convert_to_vimodule(module: nn.Module, n_args: int) -> None:
     for _ in range(n_args):
         sample.append(torch.randn(2, 3, 5))
 
-    convert_to_vimodule(module)
+    module1 = deepcopy(module)
+    convert_to_vimodule(module1)
+    ref_class = getattr(torch_blue.vi, module1.__class__.__name__)
+    assert ref_class == type(module1)
+    module1(*sample)
 
-    module(*sample)
+    convert.ban_torch_convert(module.__class__.__name__)
+    module2 = deepcopy(module)
+    convert_to_vimodule(module2)
+    assert ref_class != type(module2)
+    module2(*sample)
+
+    module3 = deepcopy(module)
+    convert_to_vimodule(module3)
+    assert type(module2) == type(module3)
+    module3(*sample)
+
+    convert.ban_reuse(module.__class__.__name__)
+    module4 = deepcopy(module)
+    convert_to_vimodule(module4)
+    assert type(module4) is not type(module3)
+    module4(*sample)
+    convert.ban_reuse(module.__class__.__name__, False)
+    convert.ban_torch_convert(module.__class__.__name__, False)
+
+    convert.ban_convert(type(module))
+    module9 = deepcopy(module)
+    with pytest.raises(
+        AttributeError,
+        match=f"'{module.__class__.__name__}' object has no attribute '_set_sampling_responsibility'",
+    ):
+        convert_to_vimodule(module9)
+    convert.ban_convert(type(module), False)
 
 
 @pytest.mark.parametrize(
@@ -52,3 +86,99 @@ def test_keep_weights(module: nn.Module, n_args: int) -> None:
         assert torch.allclose(ref[1], out[1])
     else:
         assert torch.allclose(ref, out)
+
+
+@pytest.mark.parametrize("ban_norms", [True, False])
+def test_blacklist(ban_norms: bool) -> None:
+    """Test basic blacklist and norm setting."""
+    assert torch.nn.ReLU in convert._blacklist
+    assert torch.nn.LayerNorm in convert._torch_norms
+
+    convert.convert_norms(not ban_norms)
+    if ban_norms:
+        assert convert._blacklist & convert._torch_norms == convert._torch_norms
+    else:
+        assert convert._blacklist & convert._torch_norms == set()
+
+    convert.convert_norms(True)
+    assert convert._blacklist & convert._torch_norms == set()
+
+
+@pytest.mark.parametrize("ban_norms", [True, False])
+def test_ban_convert(ban_norms: bool) -> None:
+    """Test adding and removing modules from blacklist."""
+    convert.convert_norms(not ban_norms)
+
+    module = torch.nn.Linear
+    assert module not in convert._blacklist
+    convert.ban_convert(module)
+    assert module in convert._blacklist
+    convert.ban_convert(module, False)
+    assert module not in convert._blacklist
+
+    module_list = [nn.Conv1d, nn.Conv2d, nn.Conv3d]
+    for module in module_list:
+        assert module not in convert._blacklist
+    convert.ban_convert(module_list)
+    for module in module_list:
+        assert module in convert._blacklist
+    convert.ban_convert(module_list, False)
+    for module in module_list:
[…]
