### 2025-11-09 09721c3
Add convert class reuse and ban list interces
