+++ torch_blue/vi/utils/convert.py
@@ -1,4 +1,4 @@
-from typing import Dict, Optional, Tuple, cast
+from typing import Dict, List, Optional, Tuple, Type, Union, cast
 
 import torch
 from torch import nn
@@ -10,7 +10,242 @@ from ..distributions import MeanFieldNormal
 from .common_types import VIkwargs, _dist_any_t
 from .init import fixed_
 
-_blacklist = [nn.ReLU, nn.LayerNorm]
+__all__ = [
+    "convert_to_vimodule",
+    "ban_convert",
+    "ban_torch_convert",
+    "ban_reuse",
+]
+
+# Modules in the blacklist will not be converted. The default set contains modules that
+# do not have Parameters and therefore would not change when converted. Additionally,
+# norms are assumed to be not converted by default.
+_blacklist = {
+    nn.Identity,
+    # activations
+    nn.Threshold,
+    nn.ReLU,
+    nn.RReLU,
+    nn.Hardtanh,
+    nn.ReLU6,
+    nn.Sigmoid,
+    nn.Hardsigmoid,
+    nn.Tanh,
+    nn.SiLU,
+    nn.Mish,
+    nn.Hardswish,
+    nn.ELU,
+    nn.CELU,
+    nn.SELU,
+    nn.GLU,
+    nn.GELU,
+    nn.Hardshrink,
+    nn.LeakyReLU,
+    nn.LogSigmoid,
+    nn.Softplus,
+    nn.Softshrink,
+    nn.Softsign,
+    nn.Tanhshrink,
+    nn.Softmin,
+    nn.Softmax,
+    nn.Softmax2d,
+    nn.LogSoftmax,
+    # distances
+    nn.PairwiseDistance,
+    nn.CosineSimilarity,
+    # dropout
+    nn.Dropout,
+    nn.Dropout1d,
+    nn.Dropout2d,
+    nn.Dropout3d,
+    nn.AlphaDropout,
+    nn.FeatureAlphaDropout,
+    # flatten,
+    nn.Flatten,
+    nn.Unflatten,
+    # fold
+    nn.Fold,
+    nn.Unfold,
+    # padding
+    nn.CircularPad1d,
+    nn.CircularPad2d,
+    nn.CircularPad3d,
+    nn.ConstantPad1d,
+    nn.ConstantPad2d,
+    nn.ConstantPad3d,
+    nn.ReflectionPad1d,
+    nn.ReflectionPad2d,
+    nn.ReflectionPad3d,
+    nn.ReplicationPad1d,
+    nn.ReplicationPad2d,
+    nn.ReplicationPad3d,
+    nn.ZeroPad1d,
+    nn.ZeroPad2d,
+    nn.ZeroPad3d,
+    # pixel shuffle
+    nn.PixelShuffle,
+    nn.PixelUnshuffle,
+    # pooling
+    nn.MaxPool1d,
+    nn.MaxPool2d,
+    nn.MaxPool3d,
+    nn.MaxUnpool1d,
+    nn.MaxUnpool2d,
+    nn.MaxUnpool3d,
+    nn.AvgPool1d,
+    nn.AvgPool2d,
+    nn.AvgPool3d,
+    nn.FractionalMaxPool2d,
+    nn.FractionalMaxPool3d,
+    nn.LPPool1d,
+    nn.LPPool2d,
+    nn.LPPool3d,
+    nn.AdaptiveMaxPool1d,
+    nn.AdaptiveMaxPool2d,
+    nn.AdaptiveMaxPool3d,
+    nn.AdaptiveAvgPool1d,
+    nn.AdaptiveAvgPool2d,
+    nn.AdaptiveAvgPool3d,
+    # upsampling
+    nn.Upsample,
[…]
