### 2025-11-10 ab9011f
Enable converted fastpath
