### 2025-11-10 dbc0f38
Reenable some VITransformer tests
