### 2025-11-10 5674082
Add pointer to __post_init__ in docs
