### 2025-11-14 a6be356
Move paremeter initilization to dsitribution
