### 2025-11-17 e7a4482
Add debug guide
