+++ torch_blue/vi/distributions/normal.py
@@ -1,11 +1,13 @@
-from math import exp, log
-from typing import TYPE_CHECKING, Tuple
+from math import exp, log, prod
+from typing import TYPE_CHECKING, Optional, Tuple
 
 import torch
 from torch import Tensor
+from torch.distributions.utils import lazy_property
 from torch.nn import init
 
 from torch_blue.vi import _globals
+from torch_blue.vi.utils.init import fixed_
 
 from .base import Distribution
 
@@ -52,7 +54,7 @@ class MeanFieldNormal(Distribution):
         self.eps = eps
         self._default_variational_parameters = (0.0, log(std))
 
-    @property
+    @lazy_property
     def std(self) -> float:
         """Prior standard deviation."""
         return exp(self.log_std)
@@ -72,7 +74,7 @@ class MeanFieldNormal(Distribution):
         Returns
         -------
         Tensor
-            The sampled Tensor of teh same shape as `mean`.
+            The sampled Tensor of the same shape as `mean`.
         """
         std = torch.exp(log_std)
         return self._normal_sample(mean, std)
@@ -219,3 +221,274 @@ class MeanFieldNormal(Distribution):
         if _globals._USE_NORM_CONSTANTS:
             normalization = normalization + log(2 * torch.pi)
         return -0.5 * (data_fitting + normalization)
+
+
+class CorrelatedNormal(Distribution):
+    """
+    Distribution assuming correlated, normal distributed values.
+
+    This distribution is implemented as prior, variational distribution, and predictive
+    distribution. Its distribution parameters are "mean", "log_diag", and
+    "cholesky_corr".
+
+    Parameters
+    ----------
+    mean: float, default: 0.0
+        The mean of the normal distribution before potential rescaling. Ignored if used
+        as predictive distribution.
+    std: float, default: 1.0
+        The standard deviation of the normal distribution before potential rescaling.
+        This is converted to a log std internally. For more fine-grained control,
+        specify ``scale_tril``. Ignored if used as predictive distribution or if
+        ``scale_tril`` is provided.
+    corr: float, default: 0.0
+        The correlations of the normal distribution before potential rescaling. For more
+        fine-grained control, specify ``scale_tril``. Ignored if used as predictive
+        distribution or if ``scale_tril`` is provided.
+    scale_tril: Optional[Tensor], default: None
+        If this is provided ``std`` and ``corr`` are ignored in favor of using this as
+        the Cholesky decomposition of the covariance matrix.
+    eps: float, default: 1e-10
+        Epsilon for numerical stability. Only relevant if used as prior.
+    """
+
+    # is_prior: bool = True
+    is_variational_distribution: bool = True
+    # is_predictive_distribution: bool = True
+
+    def __init__(
+        self,
+        mean: float,
+        std: float = 1.0,
+        corr: float = 0.0,
+        covariance_matrix: Optional[Tensor] = None,
+        scale_tril: Optional[Tensor] = None,
+        eps: float = 1e-10,
+    ) -> None:
+        super().__init__()
+        self.distribution_parameters = ("mean", "log_diag", "cholesky_corr")
+        self.mean = mean
+        self.log_std = log(std)
+        self.corr = corr
+        self.eps = eps
+        self._default_variational_parameters = (0.0, log(std), corr)
+
+        if (covariance_matrix is not None) + (scale_tril is not None) > 1:
+            raise ValueError(
+                "Only one of covariance_matrix or scale_tril may be specified."
+            )
+
+        if scale_tril is not None:
+            if scale_tril.dim() != 2:
+                raise ValueError("scale_tril must have two dimensions.")
+            self.scale_tril = scale_tril
[…]
