### 2025-11-21 a27fc54
Correct heat suggestions
