### 2025-11-30 RAI-SCC/torch_blue#47 @arnauqb
Issues with `pytorch_tutorial` [JOSS review]
Hello,

as part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/9415) I was having a go at the tutorial.

I am running into a few issues trying to make it work. I'm running Python 3.12.10 and torch 2.9.1.

The first issue was a shape issue with

```bash
  File "/Users/arnau/code/torch_blue/torch_blue/vi/distributions/normal.py", line 114, in _normal_sample
    base_sample = torch.normal(torch.zeros_like(mean), torch.ones_like(mean))
                  ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
RuntimeError: normal: mean and std must have same number of elements
```

which I fixed by changing that line to `base_sample = torch.randn_like(mean)`. I think it was not playing well with torch vmap.

Then it trained for a bit but had another error

```bash
now hitting Epoch 1
-------------------------------
loss: 1979169.125000  [   64/60000]
loss: 2368493.500000  [ 6464/60000]
loss: 2290490.750000  [12864/60000]
loss: 2245475.000000  [19264/60000]
loss: 2232357.250000  [25664/60000]
loss: 2111960.000000  [32064/60000]
loss: 2077849.750000  [38464/60000]
loss: 1979158.375000  [44864/60000]
loss: 1891819.625000  [51264/60000]
loss: 1901411.500000  [57664/60000]
Traceback (most recent call last):
  File "/Users/arnau/code/torch_blue/scripts/pytorch_tutorial.py", line 135, in <module>
    torch_tutorial()
  File "/Users/arnau/code/torch_blue/scripts/pytorch_tutorial.py", line 130, in torch_tutorial
    test(test_dataloader, model, loss_fn)
  File "/Users/arnau/code/torch_blue/scripts/pytorch_tutorial.py", line 119, in test
    correct += (pred.argmax(1) == y).type(torch.float).sum().item()
                ^^^^^^^^^^^^^^
  File "/Users/arnau/code/torch_blue/.venv/lib/python3.12/site-packages/torch/_tensor.py", line 1654, in
__torch_function__
    ret = func(*args, **kwargs)
          ^^^^^^^^^^^^^^^^^^^^^
IndexError: Dimension out of range (expected to be in range of [-1, 0], but got 1)
```
I changed

```bash
pred = predictive_distribution.predictive_parameters_from_samples(samples)
```

(`samples[0]` -> `samples`) and seem to work, however, the loss gets nan after at the 2nd epoch:

```bash
     Epoch 2
     -------------------------------
     loss: 1800417.125000  [   64/60000]
     loss: 1745756.750000  [ 6464/60000]
     loss: 1669817.750000  [12864/60000]
     loss:     nan  [19264/60000]
     loss:     nan  [25664/60000]
     loss:     nan  [32064/60000]
     loss:     nan  [38464/60000]
```

unsure if I broke something else with the changes. I retrained in case it was randomness issue (I would recommend fixing the seed for the tutorial) and it trained a bit more before hitting nan

```bash

Epoch 3
-------------------------------
loss: 1240194.500000  [   64/60000]
loss: 1192848.750000  [ 6464/60000]
loss: 1147005.000000  [12864/60000]
loss: 1092469.125000  [19264/60000]
loss: 1062095.000000  [25664/60000]
loss: 947314.500000  [32064/60000]
loss: 922123.500000  [38464/60000]
loss: 846004.500000  [44864/60000]
loss:     nan  [51264/60000]
loss:     nan  [57664/60000]
Test Error:
 Accuracy: 10.0%, Avg loss:      nan

Epoch 4
-------------------------------
loss:     nan  [   64/60000]
loss:     nan  [ 6464/60000]
loss:     nan  [12864/60000]
loss:     nan  [19264/60000]
loss:     nan  [25664/60000]
```
