### 2025-11-30 #9415 @arnauqb
Please see https://github.com/RAI-SCC/torch_blue/issues/47
