### 2025-12-01 c57b1d6
Implement more tests
