### 2025-12-01 9789227
Debug demo script and optimize normal distribution (#48)
Fix issues from issue (#47)
