### 2025-12-01 RAI-SCC/torch_blue#49 @JSchmie
Low accuracy compared to non-Bayesian PyTorch model
Hi,

As part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/9415), I adapted one of my existing tutorials to use `torch_blue`. Afterwards, I noticed that you use a very similar approach in your own [tutorial](https://github.com/RAI-SCC/torch_blue/blob/main/scripts/pytorch_tutorial.py). Replacing the standard PyTorch layers with your `vi` layers worked smoothly, and I was able to train the model without any issues. However, the final test performance is significantly worse than for the corresponding non-Bayesian model.

<summary>For context, my current code</summary>

```python
class simpleNN(vi.VIModule):
    def __init__(self, input_size, num_classes) -> None:
        super(simpleNN,self).__init__()
        self.fc1 = vi.VILinear(input_size,50)
        self.fc2 = vi.VILinear(50,num_classes)
    def forward(self,x):
        x = F.relu(self.fc1(x))
        x = self.fc2(x)
        return x

input_size = 784
num_classes = 10
learning_rate = 0.001
num_epochs = 25

model = simpleNN(input_size= input_size, num_classes= 10).to(device)
predictive_distribution = Categorical()
loss_fn = vi.KullbackLeiblerLoss(predictive_distribution, dataset_size=total_train_images)
optimizer = optim.Adam(model.parameters(),lr=learning_rate)

def train():
    loss_arr = []
    for epoch in trange(num_epochs):
        epoch_loss = 0
        for batch_idx,(img,label) in enumerate(trainloader):

            data = img.to(device=device)
            target = label.to(device=device)

            data = data.reshape(data.shape[0],-1) # reshape from torch.Size([64, 1, 28, 28])  torch.Size([64, 784]) which flattens the input

            score = model(data)
            loss = loss_fn(score,target)
            epoch_loss += loss.item()

            optimizer.zero_grad()
            loss.backward()

            optimizer.step()
        avg_epoch_loss = epoch_loss / len(trainloader)
        loss_arr.append(avg_epoch_loss)
        print(f"Epoch {epoch+1}/{num_epochs}, Loss: {avg_epoch_loss:.4f}")
    plt.plot(loss_arr)

train()

def check_accuracy(loader,model):
    size= len(loader.dataset)
    num_batches = len(loader)
    num_correct = 0.0
    test_loss = 0.0
    model.eval()

    with torch.no_grad():
        for x,y in loader:
            x = x.to(device = device)
            y = y.to(device = device)

            x = x.reshape(x.shape[0],-1)

            scores = model(x)

            test_loss += loss_fn(scores, y).item()

            predictions = predictive_distribution.predictive_parameters_from_samples(
                    scores[0]
                )

            num_correct += (predictions.argmax(0) == y).type(torch.float).sum().item()
    num_correct /= size
    test_loss /= num_batches
    print(
        f"Test Error: \n Accuracy: {(100 * num_correct):>0.1f}%, Avg loss: {test_loss:>8f} \n"
    )
    model.train()

check_accuracy(trainloader, model)
check_accuracy(testloader, model)
```

`scores` had the size `torch.Size([10, 64, 10])`

The Accuracy for the test set is 14.8% using torch blue vs 86.76% using the non-Bayesian model.
Could you tell me what I did wrong or why the accuracy is so low?
