### 2025-12-02 RAI-SCC/torch_blue#49 @ArvidWeyrauch
Hi,

thank you for trying things. Unfortunately, it is likely the answer to your question is that VI BNNs can be a little tricky to get right. I am setting up a debugging guide in the documentation that gives some common ways to approach this, which can be found [here](https://github.com/RAI-SCC/torch_blue/blob/docs_patch/docs/source/debug.rst) (still on a different branch).

Most importantly, all torch_blue losses have the `track` argument, which can track the terms of the loss separately. Plotting the components (and comparing them) can often help pinpoint what is going wrong. `data_fitting` behaves roughly like a non-Bayesian loss. If `prior_matching` is too dominant it might induce behavior that looks like overfitting. (If you post plots here I might be able to provide further insight)

As a short summary, there are three common things to do:
1. Setting the `heat` parameter to a number smaller than 1, which scales back the prior matching term. This is related to the cold posterior effect and may degrade calibration if done too aggressively. However, it sometimes stops the prior term from blurring away the training progress.
2.  Changing the initial std of the variational distribution, which can stabilize the early stages of training (both lowering or raising might be necessary, this would be a longer discussion). Learning rate warm up might also help here.
3. Changing (typically lowering) the prior std, as it might pull the weights to unreasonably large stds.

I am working on adding tools and guidance to help with this, so I am happy for any feedback on what users have trouble with and also if you find new solutions.
