### 2025-12-02 af8e6a2
modify conf.py to automatically filter README.md into a version that works with the MyST parser
