+++ docs/source/installation.rst
@@ -3,7 +3,7 @@
 Installation
 ============
 
-.. include:: ../../README.md
+.. include:: ../../README_rtd.md
    :parser: myst
    :start-after: ## Installation
    :end-before: ## Documentation
