### 2025-12-02 2a3af63
Rework docs intro files and some documentation tweaks (#41)
* Rework docs intro files and some documentation tweaks

* Add debug guide

* modify conf.py to automatically filter README.md into a version that works with the MyST parser

---------

Co-authored-by: lars <lars.heyen@kit.edu>
