+++ docs/source/installation.rst
@@ -1,0 +1,9 @@
+.. _installation:
+
+Installation
+============
+
+.. include:: ../../README_rtd.md
+   :parser: myst
+   :start-after: ## Installation
+   :end-before: ## Documentation
