+++ docs/source/_templates/index.rst
@@ -1,0 +1,11 @@
+API Reference
+=============
+
+This page contains autogenerated API reference documentation [1]_.
+
+.. toctree::
+   :maxdepth: 6
+
+   torch_blue/index
+
+.. [1] Created with `sphinx-autoapi <https://github.com/readthedocs/sphinx-autoapi>`_
