### 2025-12-04 RAI-SCC/torch_blue#49 @JSchmie
Thank you; I set it now and got an accuracy of 87.3% with predictions.argmax(-1).
A suggestion: as part of your ReadTheDocs, you could include a demo Jupyter file with explanations.
