### 2025-12-04 c0fcb81
Update and extend test of VIModule
