### 2025-12-04 9139c36
Update and extend test of VIModule
