+++ tests/test_vi/test_base.py
@@ -1,443 +1,504 @@
 import math
-from typing import Any, Dict, Optional, Tuple, Union, cast
-from warnings import filterwarnings
+from itertools import product
+from typing import Dict, Optional, Tuple, Type, cast
 
 import pytest
 import torch
 from torch import Tensor
 from torch._C._functorch import get_unwrapped
-from torch.nn import Module
 
 from torch_blue.vi import VIModule, VIReturn
 from torch_blue.vi.distributions import Distribution
 from torch_blue.vi.utils import NoVariablesError, UnsupportedDistributionError
 
 
-def test_expand_to_samples(device: torch.device) -> None:
-    """Test _expand_to_samples."""
-    shape1 = (3, 4)
-    sample1 = torch.randn(shape1, device=device)
-    out1 = VIModule._expand_to_samples(sample1, samples=5)
-    assert out1.shape == (5,) + shape1
-    for s in out1:
-        assert (s == sample1).all()
+class TestVIModule:
+    """Test core functionality of VIModule."""
 
-    shape2 = (5,)
-    sample2 = torch.randn(shape2, device=device)
-    out2 = VIModule._expand_to_samples(sample2, samples=1)
-    assert out2.shape == (1,) + shape2
-    for s in out2:
-        assert (s == sample2).all()
-
-    out3 = VIModule._expand_to_samples(None, samples=2)
-    assert (out3 == torch.tensor([False, False])).all()
-
-
-def test_no_forward_error() -> None:
-    """Test that forward throws error if not implemented."""
-    module = VIModule()
-    with pytest.raises(
-        NotImplementedError,
-        match=r'Module \[VIModule\] is missing the required "forward" function',
-    ):
-        module.forward(torch.randn((3, 4)))
-
-
-def test_sampled_forward(device: torch.device) -> None:
-    """Test _sampled_forward."""
-
-    class Test(VIModule):
-        def __init__(self, ref: Tensor) -> None:
-            super().__init__()
-            self.ref = ref
-            self._log_probs = dict(all=[])
-
-        def forward(self, x: Tensor) -> Tensor:
-            assert x.shape == self.ref.shape
-            self._log_probs["all"].append(
-                get_unwrapped(torch.randn(2, device=x.device))
-            )
-            return x - self.ref
-
-    shape1 = (3, 4)
-    sample1 = torch.randn(shape1, device=device)
-    test1 = Test(ref=sample1)
-    out = cast(VIReturn, test1.sampled_forward(sample1, samples=10))
-    lps = cast(Tensor, out.log_probs)
-    assert hasattr(out, "log_probs")
-    assert isinstance(out, VIReturn)
-    assert lps.shape == (10, 2)
-    assert torch.allclose(out, torch.zeros((10,) + shape1, device=device))
-    for r in lps[1:]:
-        assert not torch.allclose(lps[0], r)
-
-    shape2 = (5,)
-    sample2 = torch.randn(shape2, device=device)
-    test2 = Test(ref=sample2)
-    assert torch.allclose(
-        test2.sampled_forward(sample2, samples=1)[0],
-        torch.zeros((1,) + shape2, device=device),
+    @pytest.fixture(
+        params=[(("mean", "std"), (0.0, 1.0)), (("mean", "std"), (1.0, 0.3))]
     )
+    def dummy_prior(self, request: pytest.FixtureRequest) -> Type[Distribution]:
+        """Create dummy prior for testing."""
+        dist_params, default_params = request.param
 
+        class TestPrior(Distribution):
+            is_prior: bool = True
+            distribution_parameters: Tuple[str, ...] = dist_params
+            _scaling_parameters: Tuple[str, ...] = dist_params
+            mean: float = default_params[0]
[…]
