### 2025-12-05 08233f5
Add jupyter version of tutorial example
