### 2025-12-15 RAI-SCC/torch_blue#51 @JSchmie
Thank you, this look s fine to me!
