### 2026-01-03 b831766
Fix warning in PyTest raises.
