+++ tests/test_vi/test_analytical_kl_loss.py
@@ -619,7 +619,7 @@ def test_init(
                 f" {var_dist.__class__.__name__}.",
             ),
             (ValueError, "Provided model is not bayesian."),
-            (UnsupportedDistributionError, ""),
+            (UnsupportedDistributionError, "UniformPrior does not support use as predictive distribution"),
         ]
         error, message = error_list[expected_error]
         with pytest.raises(error, match=message):
