### 2026-01-03 #9415 @Bobby-Huggins
Hi all, sorry again for the delay. I only have some minor suggestions/notes.
1. The sentence below (lines 41-44 in the PDF) is a bit clunky --- I think it could use a re-wording, but I also don't understand what it's trying to say. Can you clarify?
    > The ability to use multiple independent sampling dimensions is removed, which allows to fully automate a single sampling dimensions in the outermost instance of the new base class `VIModule`, which captures the optional keyword argument `samples` specifying the number of samples.
3. I made some suggested edits to the text of the paper, in [PR #54](https://github.com/RAI-SCC/torch_blue/pull/54). These are just minor grammar tweaks and my suggestions to improve the flow of sentences --- feel free to accept or reject these as you see fit.
4. A minor change to avoid a warning in PyTest, in [PR #53](https://github.com/RAI-SCC/torch_blue/pull/53). A very minor thing, feel free to accept or reject.
5. Just something I noticed when looking at [the base distribution class](https://github.com/RAI-SCC/torch_blue/blob/dev/torch_blue/vi/distributions/base.py): Instead of the ad-hoc checks in `__post_init__`, maybe consider using a [mixin pattern](https://realpython.com/python-mixin/). E.g, you could have something like
   ```python
   from abc import ABC, abstractmethod

   class Prior(ABC):
       @property
       @abstractmethod
       def distribution_parameters(self):
           ...
       @abstractmethod
       def prior_log_prob(self, W):
           ...
       ...

   class VariationalDistribution(ABC):
        @property
        @abstractmethod
        def distribution_parameters(self):
            ...
        @property
        @abstractmethod
        def _default_variational_parameters(self):
            ...
        @abstractmethod
        def sample(self, n):
            ...
        ...
   class PredictiveDistribution(ABC):
       ...
   ```
   and so on. Then, you could write, e.g.,
   ```python
   class MeanFieldNormal(Distribution, Prior, VariationalDistribution, PredictiveDistribution):
       ...
   ```
   and it would be forced to implement the appropriate properties and methods. Just a thought for future revisions!
