### 2026-01-07 aa0dd37
Reword some sentences. (#54)
Integrate reviewer suggestions.
