+++ paper/paper.md
@@ -35,37 +35,37 @@ bibliography: paper.bib
 
 Bayesian Neural Networks (BNN) integrate uncertainty quantification in all steps of the
 training and prediction process, thereby enabling better-informed decisions
-[@arbel2023primer]. Among the different variants to implement BNNs, Variational
+[@arbel2023primer]. Among the different approaches to implementing BNNs, Variational
 Inference (VI) [@hoffman2013svi] specifically strikes a balance between the ability to
 consider a large variety of distributions while maintaining low enough compute
-requirements to potentially allow scaling to larger models.
+requirements to allow scaling to larger models.
 
 However, setting up and training BNNs is quite complicated, and existing libraries all
-either lack flexibility, scalability, or tackle Bayesian computation in general, adding
-even more complexity and therefore a huge entry barrier. Moreover, no existing framework
-directly supports direct BNN model prototyping by offering pre-programmed Bayesian
+either lack flexibility, lack scalability, or tackle Bayesian computation in general, adding
+even more complexity and therefore a huge barrier to entry. Moreover, no existing framework
+directly supports straightforward BNN model prototyping by offering pre-programmed Bayesian
 network layer types, similar to PyTorch's `nn` module. This forces any BNNs to be
 implemented from scratch, which can be challenging even for non-Bayesian networks.
 
 `torch_blue` addresses this by providing an interface that is almost identical to
-the widely used PyTorch [@ansel2024pytorch] for basic use, providing a low entry
-barrier, as well as an advanced interface designed for exploration and research.
+the widely used PyTorch [@ansel2024pytorch] for basic use, providing a low
+barrier to entry, as well as an advanced interface designed for exploration and research.
 Overall, this allows users to set up models and even custom layers without worrying
 about the Bayesian intricacies under the hood.
 
 # Statement of need
 
-To represent uncertainty BNNs do not consider their weights as point values, but
+To represent uncertainty, BNNs do not consider their weights as point values, but
 random variables, i.e., distributions. The optimization goal becomes adapting the weight
 distributions to minimize their distance to the true distribution. This requires two
 assumptions. For one, the distance between distributions needs to be defined, for
 which the Kullback-Leibler divergence [@kullback1951information] is typically used.
 Secondly, optimizing an object as complex as a distribution is a non-trivial task.
-To overcome this, VI utilizes a parametrized distribution and optimizing its parameters.
+To overcome this, VI specifies a parametrized distribution and optimizes its parameters.
 Thus, the Kullback-Leibler criterion can be simplified to the ELBO
 (**E**vidence **L**ower **BO**und) loss [@jordan1999introduction]:
 $$\mathrm{ELBO} = \mathbb{E}_{W\sim q}[\underbrace{\log p(Y|X, W)}_\mathrm{Data~fitting} - \underbrace{(\log q(W|\lambda) - \log p(W))}_\mathrm{Prior~matching}] \quad ,$$
-where $(X, Y)$ are the training inputs and labels, $W$ are the network weights, $q$ the
+where $(X, Y)$ are the training inputs and labels, $W$ the network weights, $q$ the
 variational distribution and $\lambda$ its current best fit parameters.
 
 While interest in uncertainty quantification and BNNs has been growing, support for
@@ -80,13 +80,13 @@ learning users.
 `torch_blue`. The highlight colors relate user-facing components to their position
 in \autoref{design_graph}. \label{code}](code_example.png)
 
-![Design graph of `torch_blue` colored highlights correspond to their practical
+![Design graph of `torch_blue`. Colored highlights correspond to their practical
 applications in the code example (\autoref{code}). \label{design_graph}](design_graph.png)
 
[…]
