### 2026-01-15 #9415 @ArvidWeyrauch
@JSchmie Is there a problem with the references in the paper? (as this is the only checkmark you have not set)
