### 2026-01-15 63b2b5b
Clarify unclear sentence (#55)
Also clean up line lengths
