### 2026-01-16 1e165cf
Make final publication preparations for JOSS
Add ORCIDs, add Badges, bump version to 0.9.1
