+++ README.md
@@ -1,4 +1,12 @@
-# torch_blue - A PyTorch-like framework for Bayesian learning and uncertainty estimation
+# torch_blue - A PyTorch-like library for Bayesian learning and uncertainty estimation
+
+[![status](https://img.shields.io/badge/Python-3.9+-blue.svg)](https://www.python.org/downloads/)
+[![status](https://img.shields.io/badge/License-BSD_3--Clause-blue.svg)](https://opensource.org/licenses/BSD-3-Clause)
+[![status](https://img.shields.io/pypi/v/torch-blue)](https://pypi.org/project/torch-blue/)
+[![status](https://readthedocs.org/projects/torch-blue/badge/?version=latest)](https://torch-blue.readthedocs.io/en/latest/?badge=latest)
+[![status](https://joss.theoj.org/papers/68b05d930d43e44aac0675c5bb3aade2/status.svg)](https://joss.theoj.org/papers/68b05d930d43e44aac0675c5bb3aade2)
+
+----------------------------------------------------------------------------------------
 
 `torch_blue` provides a simple way for non-expert users to implement and train Bayesian
 Neural Networks (BNNs). Currently, it only supports Variational Inference (VI), but will
