+++ docs/source/index.rst
@@ -1,7 +1,20 @@
-torch_blue -- A PyTorch-based framework for Bayesian learning and uncertainty estimation
-========================================================================================
+torch_blue -- A PyTorch-based library for Bayesian learning and uncertainty estimation
+======================================================================================
 
-Release: |release|
+| |python-version| |license| |docs| |pypi-version| |joss|
+
+.. |pypi-version| image:: https://img.shields.io/pypi/v/torch-blue
+   :target: https://pypi.org/project/torch-blue/
+.. |python-version| image:: https://img.shields.io/badge/Python-3.9+-blue.svg
+   :target: https://www.python.org/downloads/
+.. |license| image:: https://img.shields.io/badge/License-BSD_3--Clause-blue.svg
+   :target: https://opensource.org/licenses/BSD-3-Clause
+.. |docs| image:: https://readthedocs.org/projects/torch-blue/badge/?version=latest
+   :target: https://torch-blue.readthedocs.io/en/latest/?badge=latest
+.. |joss| image:: https://joss.theoj.org/papers/68b05d930d43e44aac0675c5bb3aade2/status.svg
+   :target: https://joss.theoj.org/papers/68b05d930d43e44aac0675c5bb3aade2
+
+========================================================================================
 
 ``torch_blue`` provides a simple way for non-expert users to implement and train
 Bayesian Neural Networks (BNNs). Currently, it only supports Variational Inference (VI),
