### 2026-01-19 6b3e3c6
Final preparations for JOSS release (#57)
* Make final publication preparations for JOSS

* Add ORCIDs, add Badges, bump version to 0.9.1

---------

Co-authored-by: Asena Karolin Özdemir <60694557+AsenaKarolin@users.noreply.github.com>
