### 2026-01-19 47b9838
Copy-editing; add missing DOIs
