+++ paper/paper.bib
@@ -10,8 +10,9 @@
                Paul Horsfall and
                Noah D. Goodman},
   title     = {Pyro: Deep Universal Probabilistic Programming},
-  journal   = {J. Mach. Learn. Res.},
+  journal   = {Journal of Machine Learning Research},
   volume    = {20},
+  number    = 28,
   pages     = {28:1--28:6},
   year      = {2019},
   url       = {http://jmlr.org/papers/v20/18-403.html}
@@ -26,7 +27,7 @@
 
 @inproceedings{ansel2024pytorch,
   author = {Ansel, Jason and Yang, Edward and He, Horace and Gimelshein, Natalia and Jain, Animesh and Voznesensky, Michael and Bao, Bin and Bell, Peter and Berard, David and Burovski, Evgeni and Chauhan, Geeta and Chourdia, Anjali and Constable, Will and Desmaison, Alban and DeVito, Zachary and Ellison, Elias and Feng, Will and Gong, Jiong and Gschwind, Michael and Hirsh, Brian and Huang, Sherlock and Kalambarkar, Kshiteej and Kirsch, Laurent and Lazos, Michael and Lezcano, Mario and Liang, Yanbo and Liang, Jason and Lu, Yinghai and Luk, C. K. and Maher, Bert and Pan, Yunjie and Puhrsch, Christian and Reso, Matthias and Saroufim, Mark and Siraichi, Marcos Yukio and Suk, Helen and Zhang, Shunting and Suo, Michael and Tillet, Phil and Zhao, Xu and Wang, Eikan and Zhou, Keren and Zou, Richard and Wang, Xiaodong and Mathews, Ajit and Wen, William and Chanan, Gregory and Wu, Peng and Chintala, Soumith},
-  title = {PyTorch 2: Faster Machine Learning Through Dynamic Python Bytecode Transformation and Graph Compilation},
+  title = {PyTorch~2: Faster Machine Learning Through Dynamic {Python} Bytecode Transformation and Graph Compilation},
   year = {2024},
   isbn = {9798400703850},
   publisher = {Association for Computing Machinery},
@@ -41,7 +42,6 @@
   series = {ASPLOS '24}
 }
 
-
 @InProceedings{blundell15bbb,
   title = 	 {Weight Uncertainty in Neural Network},
   author = 	 {Blundell, Charles and Cornebise, Julien and Kavukcuoglu, Koray and Wierstra, Daan},
@@ -52,7 +52,7 @@
   volume = 	 {37},
   series = 	 {Proceedings of Machine Learning Research},
   address = 	 {Lille, France},
-  month = 	 {07--09 Jul},
+  month = 	 jul,
   publisher =    {PMLR},
   pdf = 	 {http://proceedings.mlr.press/v37/blundell15.pdf},
   url = 	 {https://proceedings.mlr.press/v37/blundell15.html},
@@ -61,23 +61,26 @@
 
 @article{kullback1951information,
   title={On information and sufficiency},
-  author={Kullback, Solomon and Leibler, Richard A},
-  journal={The annals of mathematical statistics},
+  author={Kullback, Solomon and Leibler, Richard A.},
+  journal={The Annals of Mathematical Statistics},
   volume={22},
   number={1},
   pages={79--86},
   year={1951},
-  publisher={JSTOR}
+  month=mar,
+  doi={10.1214/aoms/1177729694},
 }
 
 @article{jordan1999introduction,
   title={An introduction to variational methods for graphical models},
-  author={Jordan, Michael I and Ghahramani, Zoubin and Jaakkola, Tommi S and Saul, Lawrence K},
-  journal={Machine learning},
[…]
