### 2026-01-19 #9415 @logological
@ArvidWeyrauch I have submitted pull request RAI-SCC/torch_blue#59 to correct some minor language issues and to add some missing DOIs.  Could you please review this and, if you are happy, accept the changes?
