### 2026-01-19 #9415 @logological
@ArvidWeyrauch In case you do accept the pull request as-is, then we can move on to some final tasks.  In this case, could you please do the following?

1. Double-check that the paper's list of authors, including affiliations and ORCIDs, is correct.
2. Make a new release of the software incorporating the latest changes arising from the review and post the version number here. This is the version that will be used in accepted version of the JOSS paper.
3. Archive the software release on Zenodo (or figshare or some similar archiving service) making sure that
    - the title and author list (including ORCIDs) given in the archiving service exactly matches the title and author list in the paper; and
    - the software licence given in the archiving service exactly matches the software licence specified in your source repository.
4. Post the DOI of your archive here.
