### 2026-01-20 #9415 @ArvidWeyrauch
@logological Thank you for the fixes. I have merged the them as-is.
Step 1. is already done. I am in the process of making a PyPI release for the submission version, the version number will be 0.9.1, but this might take a few more days. I will let you know when that is done and uploaded to Zenodo.

Right at the start of the pre-review the editor in chief mention that I should [remove the quotes from the title](https://github.com/openjournals/joss-reviews/issues/8968#issuecomment-3276598415). I am still not sure what he meant with that, since the only title I am able to change is the one in the paper header, where the sample paper also uses quotes. Do you have insights there and is this still an issue?
