### 2026-01-20 06236ab
Beta 0.9.1
This is the release associated with the JOSS publication of torch_blue

* Add JOSS paper and workflow

*  Upgrade and clean README and docs

* Add Debug Guide to docs

* Add Contribution guidelines

* Debug MNIST Tutorial script and optimize normal distribution (#48)

* Add Jupyter version of tutorial example (#52)

* Integrate several reviewer suggestions and fixes

---------

Co-authored-by: lars <lars.heyen@kit.edu>
Co-authored-by: Asena Karolin Özdemir <60694557+AsenaKarolin@users.noreply.github.com>

Special thanks for the contributions from our reviewers and editor
Co-authored
[…]
