+++ paper/paper.md
@@ -1,0 +1,174 @@
+---
+title: 'torch_blue: A Flexible Python Package for Bayesian Neural Networks in PyTorch'
+tags:
+    - Python
+    - Bayesian Neural Networks
+    - Variational Inference
+authors:
+    - name: Arvid Weyrauch
+      orcid: 0000-0002-2684-0927
+      affiliation: 1
+    - name: Lars H. Heyen
+      orcid: 0000-0001-7949-1858
+      affiliation: 1
+    - name: Juan Pedro Gutiérrez Hermosillo Muriedas
+      orcid: 0000-0001-8439-7145
+      affiliation: 1
+    - name: Pei-Hsuan Hsia
+      orcid: 0009-0003-5440-4211
+      affiliation: 1
+    - name: Asena Karolin Özdemir
+      orcid: 0009-0003-2808-4744
+      affiliation: 1
+    - name: Achim Streit
+      orcid: 0000-0002-5065-469X
+      affiliation: 1
+    - name: Markus Götz
+      orcid: 0000-0002-2233-1041
+      affiliation: "1, 2"
+    - name: Charlotte Debus
+      orcid: 0000-0002-7156-2022
+      affiliation: 1
+affiliations:
+    - name: Karlsruhe Institute of Technology, Germany
+      index: 1
+      ror: 04t3en479
+    - name: Helmholtz AI
+      index: 2
+date: 16 January 2026
+bibliography: paper.bib
+---
+
+# Summary
+
+Bayesian Neural Networks (BNN) integrate uncertainty quantification in all steps of the
+training and prediction process, thereby enabling better-informed decisions
+[@arbel2023primer]. Among the different approaches to implementing BNNs, Variational
+Inference (VI) [@hoffman2013svi] specifically strikes a balance between the ability to
+consider a large variety of distributions while maintaining low enough compute
+requirements to allow scaling to larger models.
+
+However, setting up and training BNNs is quite complicated, and existing libraries all
+either lack flexibility, lack scalability, or tackle Bayesian computation in general,
+adding even more complexity and therefore a huge barrier to entry. Moreover, no existing
+framework directly supports straightforward BNN model prototyping by offering
+pre-programmed Bayesian network layer types, similar to PyTorch's `nn` module. This
+forces any BNNs to be implemented from scratch, which can be challenging even for
+non-Bayesian networks.
+
+`torch_blue` addresses this by providing an interface that is almost identical to
+the widely used PyTorch [@ansel2024pytorch] for basic use, providing a low
+barrier to entry, as well as an advanced interface designed for exploration and research.
+Overall, this allows users to set up models and even custom layers without worrying
+about the Bayesian intricacies under the hood.
+
+# Statement of need
+
+To represent uncertainty, BNNs do not consider their weights as point values, but
+random variables, i.e., distributions. The optimization goal becomes adapting the weight
+distributions to minimize their distance to the true distribution. This requires two
+assumptions. For one, the distance between distributions needs to be defined, for
+which the Kullback-Leibler divergence [@kullback1951information] is typically used.
+Secondly, optimizing an object as complex as a distribution is a non-trivial task.
+To overcome this, VI specifies a parametrized distribution and optimizes its parameters.
+Thus, the Kullback-Leibler criterion can be simplified to the ELBO
+(**E**vidence **L**ower **BO**und) loss [@jordan1999introduction]:
+$$\mathrm{ELBO} = \mathbb{E}_{W\sim q}[\underbrace{\log p(Y|X, W)}_\mathrm{Data~fitting} - \underbrace{(\log q(W|\lambda) - \log p(W))}_\mathrm{Prior~matching}] \quad ,$$
+where $(X, Y)$ are the training inputs and labels, $W$ the network weights, $q$ the
[…]
