+++ docs/source/introduction.rst
@@ -1,6 +1,0 @@
-Introduction
-============
-
-.. include:: ../../README.md
-   :parser: myst
-   :end-before: ## Installation
