+++ pyproject.toml
@@ -12,7 +12,7 @@ packages = [
 
 [project]
 name = "torch_blue"
-version = "0.9.0"
+version = "0.9.1"
 authors = [
   { name="RAI", email="rai@kit.edu"},
 ]

[…]
