### 2026-01-20 #9415 @ArvidWeyrauch
@logological things went more smoothly than expected. There is now a release of the current state on PyPI and archived on Zenodo:
https://doi.org/10.5281/zenodo.18313585

I have double checked the author list, title and license. Let me know if anything is still missing.
