### 2026-01-20 #9415 @logological
> Right at the start of the pre-review the editor in chief mention that I should [remove the quotes from the title](https://github.com/openjournals/joss-reviews/issues/8968#issuecomment-3276598415). I am still not sure what he meant with that, since the only title I am able to change is the one in the paper header, where the sample paper also uses quotes. Do you have insights there and is this still an issue?

No, I'm not sure what that's referring to.  As far as I know, the quotes are actually required by YAML syntax (since your paper title contains a colon), and as you noted, the example paper from the documentation also wraps its title in quotes.  I think we can ignore this for now; if the AEiC has a problem with how the title is formatted, then they can let us know after I've recommended acceptance.
