### 2026-01-21 ac82e24
Forward reroute (#43)
* Implement forward rerouting to avoid code duplication from PyTorch
