+++ torch_blue/vi/base.py
@@ -1,21 +1,11 @@
-import warnings
 from copy import deepcopy
 from typing import Any, Callable, Dict, List, Optional, Tuple, Union, cast
 
 import torch
-import torch.utils.hooks as hooks
 from torch import Tensor
 from torch._C._functorch import get_unwrapped
 from torch.nn import Module, Parameter, init
 from torch.nn.common_types import _tensor_list_t
-from torch.nn.modules.module import (
-    _global_backward_hooks,
-    _global_backward_pre_hooks,
-    _global_forward_hooks,
-    _global_forward_hooks_always_called,
-    _global_forward_pre_hooks,
-    _WrappedHook,
-)
 
 from .distributions import MeanFieldNormal
 from .utils import NoVariablesError, PostInitCallMeta, UnsupportedDistributionError
@@ -131,6 +121,7 @@ class VIModule(Module, metaclass=PostInitCallMeta):
     # and the outermost module calls sampled_forward instead
     _has_sampling_responsibility: bool
     _log_probs: Dict[str, List[Tensor]]
+    _forward_rerouted: bool = False
 
     def __init__(
         self,
@@ -418,6 +409,11 @@ class VIModule(Module, metaclass=PostInitCallMeta):
         and copies the input batch the specified number of times. The :meth:`~forward`
         is performed vectorized over that additional sample dimension.
 
+        If you need more information on how this is achieved, check the documentation of
+        this class' `__post_init__` method in the source code. Note this goes in deep
+        and if checking the source code seems too much hassle you should probably not
+        touch it and write an issue instead.
+
         Parameters
         ----------
         input_: Tensor
@@ -433,7 +429,7 @@ class VIModule(Module, metaclass=PostInitCallMeta):
             One or multiple Tensors with log prob annotation
         """
         expanded = [self._expand_to_samples(x, samples=samples) for x in input_]
-        out: _tensor_list_t = torch.vmap(self.forward, randomness="different")(
+        out: _tensor_list_t = torch.vmap(self._module_forward, randomness="different")(
             *expanded, **kwargs
         )
 
@@ -509,6 +505,15 @@ class VIModule(Module, metaclass=PostInitCallMeta):
                 module._has_sampling_responsibility = False
         self._has_sampling_responsibility = True
 
+    def _reroute_forward(
+        self, new_forward: Callable, new_name: str = "_module_forward"
+    ) -> None:
+        if self._forward_rerouted:
+            return
+        setattr(self, new_name, self.forward)
+        self.forward = new_forward
+        self._forward_rerouted = True
+
     def __post_init__(self) -> None:
         """
         After __init__ set sampling responsibility.
@@ -516,198 +521,23 @@ class VIModule(Module, metaclass=PostInitCallMeta):
         Since higher level modules overwrite _has_sampling_responsibility for lower ones
         only the top level class will have it set to True, making it use sampled_forward
         by default.
+        Additionally, the module defined forward method is reassigned to the name
+        _module_forward and self._pre_forward takes its place. This creates a wrapper
+        and hook just around the forward method to allow handling of automatic sampling
+        and log probability tracking.
+        If you wish to modify this behavior try using PyTorch's forward hooks first. If
+        those prove insufficient, you can modify the _pre_forward method, but be aware
+        this might break autosampling and log prob tracking if done incorrectly.
         """
         self._set_sampling_responsibility()
+        self._reroute_forward(self._pre_forward)
 
     def _pre_forward(self, *input_: Any, **kwargs: Any) -> Any:
         """Select sampled_forward or forward based on _has_sampling_responsibility."""
         if self._has_sampling_responsibility:
[…]
