### 2026-01-21 3de363a
Add codecov badges to README and docs
