### 2026-01-22 2971c69
Add epsilon to MeanFieldNormal.variational_log_prob
