### 2026-01-22 a480658
Switch to recursive conversion approach
This allows to ban conversion of all submodules of a specified
module which is also implemented.
Furthermore, all banlist modifications (except bulk en- and
disabling of norms) are joined in one method.
