+++ torch_blue/vi/utils/convert.py
@@ -1,4 +1,4 @@
-from typing import Dict, List, Optional, Tuple, Type, Union, cast
+from typing import Dict, List, Optional, Set, Tuple, Type, Union, cast
 
 import torch
 from torch import nn
@@ -11,10 +11,9 @@ from .common_types import VIkwargs, _dist_any_t
 from .init import fixed_
 
 __all__ = [
+    "convert_norms",
     "convert_to_vimodule",
     "ban_convert",
-    "ban_torch_convert",
-    "ban_reuse",
 ]
 
 # Modules in the blacklist will not be converted. The default set contains modules that
@@ -133,8 +132,9 @@ _torch_norms = {
     nn.RMSNorm,
 }
 _blacklist |= _torch_norms
-_torch_blacklist = set()
-_reuse_blacklist = set()
+_replace_blacklist: Set[nn.Module] = set()
+_reuse_blacklist: Set[nn.Module] = set()
+_submodule_blacklist: Set[nn.Module] = set()
 
 
 def convert_norms(mode: bool = True) -> None:
@@ -164,88 +164,59 @@ def convert_norms(mode: bool = True) -> None:
 
 
 def ban_convert(
-    class_names: Union[Type[nn.Module], List[Type[nn.Module]]], mode: bool = True
+    class_names: Union[Type[nn.Module], List[Type[nn.Module]]],
+    ban_mode: str = "ban",
+    unban: bool = False,
 ) -> None:
     """
-    Ban given class names from conversions.
+    Ban given class or list of classes from conversions.
 
-    This method adds one or more class names to the conversion blacklist so they are not
-    touched during auto-conversion keeping them non-Bayesian. The default blacklist
-    contains a variety to PyTorch modules that do not have parameters and therefore
-    would not change during conversion. Additionally, by default PyTorch normalization
-    layers are not converted, since their purpose is only stability not learning. If you
-    wish to change this behavior use :func:`~convert_norms` to add or remove all PyTorch
-    norms from the blacklist.
+    This method adds one or more class names to one of the conversion blacklists. There
+    are several different ban modes as defined by the `ban_mode` argument.
+
+    The default blacklist contains a variety to PyTorch modules that do not have
+    parameters and therefore would not change during conversion. Additionally, by
+    default PyTorch normalization layers are not converted, since their purpose is only
+    stability not learning. If you wish to change this behavior use
+    :func:`~convert_norms` to add or remove all PyTorch norms from the blacklist.
 
     Parameters
     ----------
-    class_names: Union[str, List[str]]
-        A class name or a list of class names not to be converted.
-    mode: bool, default=True
-        If `False` this will unban the provided names instead.
+    class_names: Union[nn.Module, List[nn.Module]]
+        A module class or a list of module classes to add to a blacklist.
+    ban_mode: str, default="ban"
+        "ban": The module is not changed by auto-conversion, but submodules are still
+        converted.
+        "submodule": Both the module and all its submodules are not converted.
+        "reuse": The module is converted, but if it occurs multiple times the
+        auto-generated class will not be reused.
+        "replace": The module will be converted and an auto-generated class is created
+        even if `torch_blue` implements a module with the same name.
+    unban: bool, default=False
+        If `True` this will unban the provided classes instead.
     """
-    global _torch_blacklist
+    if ban_mode == "ban":
+        global _blacklist
+        blacklist = _blacklist
+    elif ban_mode == "reuse":
+        global _reuse_blacklist
+        blacklist = _reuse_blacklist
+    elif ban_mode == "replace":
+        global _replace_blacklist
+        blacklist = _replace_blacklist
+    elif ban_mode == "submodule":
+        global _submodule_blacklist
+        blacklist = _submodule_blacklist
+    else:
[…]
