### 2026-01-23 #9415 @crvernon
🥳 Congratulations on your new publication @ArvidWeyrauch ! Many thanks to @logological for editing and @Bobby-Huggins, @arnauqb, and @JSchmie for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
